% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admiral_ae}
\alias{admiral_ae}
\title{Adverse Events Dataset-updated}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1191 rows and 36 columns.
}
\source{
\url{https://github.com/pharmaverse/admiral.test/blob/main/data/admiral_ae.rda}
}
\usage{
admiral_ae
}
\description{
An updated SDTM AE dataset that uses the CDISC pilot project
}
\keyword{datasets}
