% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admiral_ds}
\alias{admiral_ds}
\title{Disposition Dataset-updated}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 850 rows and 13 columns.
}
\source{
\url{https://github.com/pharmaverse/admiral.test/blob/main/data/admiral_ds.rda}
}
\usage{
admiral_ds
}
\description{
An updated SDTM DS dataset that uses the CDISC pilot project
}
\author{
Gopi Vegesna
}
\keyword{datasets}
