% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admiral_lb}
\alias{admiral_lb}
\title{Laboratory Measurements Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 59580 rows and 23 columns.
}
\source{
\url{https://github.com/pharmaverse/admiral.test/blob/main/data/admiral_lb.rda}
}
\usage{
admiral_lb
}
\description{
An updated SDTM LB dataset that uses data from the CDISC pilot project
}
\author{
Annie Yang
}
\keyword{datasets}
