% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admiral_pp}
\alias{admiral_pp}
\title{Pharmacokinetic Parameters Dataset}
\format{
An object of class \code{data.frame} with 2346 rows and 14 columns.
}
\source{
\url{https://github.com/pharmaverse/admiral.test/blob/main/data/admiral_pp.rda}
}
\usage{
admiral_pp
}
\description{
A SDTM PP dataset simulated by Antonio Rodríguez Contestí
}
\author{
Antonio Rodríguez Contestí
}
\keyword{datasets}
