% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admiral_qs}
\alias{admiral_qs}
\title{Questionnaire Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 202907 rows and 20 columns.
}
\source{
\url{https://github.com/pharmaverse/admiral.test/blob/main/data/admiral_qs.rda} # nolint
}
\usage{
admiral_qs
}
\description{
A SDTM QS dataset from the CDISC pilot project & Ophthalmology test data
}
\keyword{datasets}
