% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admiral_sc}
\alias{admiral_sc}
\title{Subject Characteristic Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 254 rows and 12 columns.
}
\usage{
admiral_sc
}
\description{
A SDTM SC dataset simulated by Ophthalmology team
}
\author{
Gordon Miller
}
\keyword{datasets}
