% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admiral_tr}
\alias{admiral_tr}
\title{Tumor Results Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 55995 rows and 24 columns.
}
\usage{
admiral_tr
}
\description{
A SDTM TR dataset simulated by Gopi Vegesna
}
\author{
Gopi Vegesna
}
\keyword{datasets}
