% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admiraldata-package.R
\docType{package}
\name{admiraldata-package}
\alias{admiraldata}
\alias{admiraldata-package}
\title{admiraldata: Test Data for the 'admiral' Package}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A set of Study Data Tabulation Model (SDTM) datasets from the Clinical Data Interchange Standards Consortium (CDISC) pilot project used for testing and developing Analysis Data Model (ADaM) derivations inside the 'admiral' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pharmaverse.github.io/admiraldata}
  \item \url{https://github.com/pharmaverse/admiraldata}
}

}
\author{
\strong{Maintainer}: Edoardo Mancini \email{edoardo.mancini@roche.com}

}
\keyword{internal}
