% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ae_ophtha}
\alias{ae_ophtha}
\title{Adverse Events for Ophthalmology}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1191 rows and 36 columns.
}
\source{
Constructed using \code{ae} from the \code{{pharmaversesdtm}} package
}
\usage{
ae_ophtha
}
\description{
An example Adverse Events SDTM dataset with ophthalmology-specific variable \code{AELAT}
}
\keyword{datasets}
