% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_neuro.R
\docType{data}
\name{ag_neuro}
\alias{ag_neuro}
\title{Procedure Agents Dataset-neuro}
\format{
A data frame with 13 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{AGSEQ}{Sequence Number}
\item{AGTRT}{Reported Agent Name}
\item{AGCAT}{Category for Agent}
\item{AGDOSE}{Dose per Administration}
\item{AGDOSEU}{Dose Units}
\item{AGROUTE}{Route of Administration}
\item{AGLNKID}{Link ID}
\item{VISITNUM}{Visit Number}
\item{VISIT}{Visit Name}
\item{AGSTDTC}{Start Date/Time of Agent}
}
}
\source{
Constructed by {admiralneuro} developers
}
\usage{
ag_neuro
}
\description{
A SDTM AG dataset recording details of agents such as a PET tracer used in procedures
}
\details{
Procedure Agents Dataset-neuro

A SDTM AG dataset recording details of agents such as a PET tracer used in procedures
}
\keyword{dataset}
