% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/be.R
\docType{data}
\name{be}
\alias{be}
\title{Biospecimen Events}
\format{
A data frame with 13 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{BESEQ}{Sequence Number}
\item{BEREFID}{Reference ID}
\item{BELNKID}{Link Identifier}
\item{BETERM}{Reported Term for the Biospecimen Event}
\item{BECAT}{Category for Biospecimen Event}
\item{BELOC}{Anatomical Location of Event}
\item{VISITNUM}{Visit Number}
\item{BEDTC}{Date/Time of Specimen Collection}
\item{BESTDTC}{Start Date/Time of Biospecimen Event}
\item{BEENDTC}{End Date/Time of Biospecimen Event}
}
}
\source{
\href{https://github.com/pharmaverse/pharmaversesdtm/blob/main/data/be.rda}{Access the source of the Biospecimen Events dataset.}
}
\usage{
be
}
\description{
A synthetic SDTM BE domain representing specimen collection, aliquoting, and culturing events, with linkage to MB and MS domains
}
\details{
Biospecimen Events

A synthetic SDTM BE domain representing specimen collection, aliquoting, and culturing events, with linkage to MB and MS domains
}
\keyword{dataset}
\keyword{microbiology}
