% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_metabolic.R
\docType{data}
\name{dm_metabolic}
\alias{dm_metabolic}
\title{Demographic Dataset-metabolic}
\format{
A data frame with 25 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{SUBJID}{Subject Identifier for the Study}
\item{RFSTDTC}{Subject Reference Start Date/Time}
\item{RFENDTC}{Subject Reference End Date/Time}
\item{RFXSTDTC}{Date/Time of First Study Treatment}
\item{RFXENDTC}{Date/Time of Last Study Treatment}
\item{RFICDTC}{Date/Time of Informed Consent}
\item{RFPENDTC}{Date/Time of End of Participation}
\item{DTHDTC}{Date/Time of Death}
\item{DTHFL}{Subject Death Flag}
\item{SITEID}{Study Site Identifier}
\item{AGE}{Age}
\item{AGEU}{Age Units}
\item{SEX}{Sex}
\item{RACE}{Race}
\item{ETHNIC}{Ethnicity}
\item{ARMCD}{Planned Arm Code}
\item{ARM}{Description of Planned Arm}
\item{ACTARMCD}{Actual Arm Code}
\item{ACTARM}{Description of Actual Arm}
\item{COUNTRY}{Country}
\item{DMDTC}{Date/Time of Collection}
\item{DMDY}{Study Day of Collection}
}
}
\source{
Constructed by {admiralmetabolic} developers
}
\usage{
dm_metabolic
}
\description{
A SDTM DM dataset for metabolic studies
}
\details{
Demographic Dataset-metabolic

A SDTM DM dataset for metabolic studies
}
\keyword{dataset}
