% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_neuro.R
\docType{data}
\name{dm_neuro}
\alias{dm_neuro}
\title{Demographic Dataset-neuro}
\format{
A data frame with 27 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{SUBJID}{Subject Identifier for the Study}
\item{RFSTDTC}{Subject Reference Start Date/Time}
\item{RFENDTC}{Subject Reference End Date/Time}
\item{RFXSTDTC}{Date/Time of First Study Treatment}
\item{RFXENDTC}{Date/Time of Last Study Treatment}
\item{RFICDTC}{Date/Time of Informed Consent}
\item{RFPENDTC}{Date/Time of End of Participation}
\item{DTHDTC}{Date/Time of Death}
\item{DTHFL}{Subject Death Flag}
\item{SITEID}{Study Site Identifier}
\item{AGE}{Age}
\item{BRTHDTC}{Date/Time of Birth}
\item{AGEU}{Age Units}
\item{SEX}{Sex}
\item{RACE}{Race}
\item{ETHNIC}{Ethnicity}
\item{ARMCD}{Planned Arm Code}
\item{ARM}{Description of Planned Arm}
\item{ACTARMCD}{Actual Arm Code}
\item{ACTARM}{Description of Actual Arm}
\item{COUNTRY}{Country}
\item{DMDTC}{Date/Time of Collection}
\item{DMDY}{Study Day of Collection}
\item{ARMNRS}{Reason Arm and/or Actual Arm is Null}
}
}
\source{
Constructed by {admiralneuro} developers
}
\usage{
dm_neuro
}
\description{
A SDTM DM dataset for Alzheimer's disease studies
}
\details{
Demographic Dataset-neuro

A SDTM DM dataset for Alzheimer's disease studies
}
\keyword{dataset}
\keyword{neurology}
