% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/dm_peds.R
\docType{data}
\name{dm_peds}
\alias{dm_peds}
\title{Demographic Dataset-pediatrics}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 26 columns.

A data frame with 26 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{SUBJID}{Subject Identifier for the Study}
\item{RFSTDTC}{Subject Reference Start Date/Time}
\item{RFENDTC}{Subject Reference End Date/Time}
\item{RFXSTDTC}{Date/Time of First Study Treatment}
\item{RFXENDTC}{Date/Time of Last Study Treatment}
\item{RFICDTC}{Date/Time of Informed Consent}
\item{RFPENDTC}{Date/Time of End of Participation}
\item{DTHDTC}{Date/Time of Death}
\item{DTHFL}{Subject Death Flag}
\item{SITEID}{Study Site Identifier}
\item{AGE}{Age}
\item{AGEU}{Age Units}
\item{SEX}{Sex}
\item{RACE}{Race}
\item{ETHNIC}{Ethnicity}
\item{ARMCD}{Planned Arm Code}
\item{ARM}{Description of Planned Arm}
\item{ACTARMCD}{Actual Arm Code}
\item{ACTARM}{Description of Actual Arm}
\item{COUNTRY}{Country}
\item{DMDTC}{Date/Time of Collection}
\item{DMDY}{Study Day of Collection}
\item{BRTHDTC}{Date/Time of Birth}
}
}
\source{
Constructed by \code{{admiralpeds}} developers

Constructed by {admiralpeds} developers
}
\usage{
dm_peds

dm_peds
}
\description{
An updated SDTM DM dataset with pediatric patients

Demographic Dataset-pediatrics
}
\details{
An updated SDTM DM dataset with pediatric patients
}
\keyword{datasets}
