% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.R
\docType{data}
\name{ds}
\alias{ds}
\title{Disposition}
\format{
A data frame with 13 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{DSSEQ}{Sequence Number}
\item{DSSPID}{Sponsor-Defined Identifier}
\item{DSTERM}{Reported Term for the Disposition Event}
\item{DSDECOD}{Standardized Disposition Term}
\item{DSCAT}{Category for Disposition Event}
\item{VISITNUM}{Visit Number}
\item{VISIT}{Visit Name}
\item{DSDTC}{Date/Time of Collection}
\item{DSSTDTC}{Start Date/Time of Disposition Event}
\item{DSSTDY}{Study Day of Start of Disposition Event}
}
}
\source{
\href{https://github.com/pharmaverse/pharmaversesdtm/blob/main/data/ds.rda}{Access the source of the Disposition dataset.}
}
\usage{
ds
}
\description{
An updated SDTM DS dataset that uses the CDISC pilot project
}
\details{
Disposition

An updated SDTM DS dataset that uses the CDISC pilot project
}
\author{
Gopi Vegesna
}
\keyword{dataset}
