% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/ex_ophtha.R
\docType{data}
\name{ex_ophtha}
\alias{ex_ophtha}
\title{Exposure for Ophthalmology}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 591 rows and 19 columns.

A data frame with 19 columns:
\describe{
\item{USUBJID}{Unique Subject Identifier}
\item{STUDYID}{Study Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{EXSEQ}{Sequence Number}
\item{EXTRT}{Name of Actual Treatment}
\item{EXDOSE}{Dose per Administration}
\item{EXDOSU}{Dose Units}
\item{EXDOSFRM}{Dose Form}
\item{EXDOSFRQ}{Dose Frequency per Interval}
\item{EXROUTE}{Route of Administration}
\item{EXLOC}{Location of Dose Administration}
\item{EXLAT}{Laterality}
\item{VISITNUM}{Visit Number}
\item{VISIT}{Visit Name}
\item{VISITDY}{Planned Study Day of Visit}
\item{EXSTDTC}{Start Date/Time of Treatment}
\item{EXENDTC}{End Date/Time of Treatment}
\item{EXSTDY}{Study Day of Start of Treatment}
\item{EXENDY}{Study Day of End of Treatment}
}
}
\source{
Constructed using \code{ex} from the \code{{pharmaversesdtm}} package

Constructed using ex from the {pharmaversesdtm} package
}
\usage{
ex_ophtha

ex_ophtha
}
\description{
An example Exposure SDTM dataset with ophthalmology-specific variables such as \code{EXLOC} and \code{EXLAT}

Exposure for Ophthalmology
}
\details{
An example Exposure SDTM dataset with ophthalmology-specific variables such as EXLOC and EXLAT
}
\keyword{datasets}
