% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/ex_vaccine.R
\docType{data}
\name{ex_vaccine}
\alias{ex_vaccine}
\title{Exposures for Vaccine}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 21 columns.

A data frame with 21 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{EXSEQ}{Sequence Number}
\item{EXLNKGRP}{Link Group ID}
\item{EXLNKID}{Link ID}
\item{EXTRT}{Name of Actual Treatment}
\item{EXCAT}{Category of Treatment}
\item{EXDOSE}{Dose per Administration}
\item{EXDOSU}{Dose Units}
\item{EXDOSFRM}{Dose Form}
\item{EXROUTE}{Route of Administration}
\item{EXLOC}{Location of Dose Administration}
\item{EXLAT}{Laterality}
\item{VISITNUM}{Visit Number}
\item{VISIT}{Visit Name}
\item{EPOCH}{Epoch}
\item{EXDTC}{undocumented field}
\item{EXSTDTC}{Start Date/Time of Treatment}
\item{EXENDTC}{End Date/Time of Treatment}
\item{EXDY}{undocumented field}
}
}
\source{
Constructed by \code{{admiralvaccine}} developers

Constructed by {admiralvaccine} developers
}
\usage{
ex_vaccine

ex_vaccine
}
\description{
An example SDTM EX dataset for vaccine studies

Exposures for Vaccine
}
\details{
An example SDTM EX dataset for vaccine studies
}
\keyword{datasets}
