% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_terms.R
\name{get_terms}
\alias{get_terms}
\title{An example function as expected by the \code{get_terms_fun} parameter of
\code{admiral::create_query_data()}}
\usage{
get_terms(basket_select, version, keep_id, temp_env)
}
\arguments{
\item{basket_select}{A basket_select object defining the terms}

\item{version}{MedDRA version}

\item{keep_id}{Should \code{GRPID} be included in the output?}

\item{temp_env}{Temporary environment}
}
\description{
An example function as expected by the \code{get_terms_fun} parameter of
\code{admiral::create_query_data()}
}
