% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_ada.R
\docType{data}
\name{is_ada}
\alias{is_ada}
\title{Immunogenicity Specimen Assessments for ADA}
\format{
A data frame with 27 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{ISSEQ}{Sequence Number}
\item{ISTESTCD}{Immunogenicity Test/Exam Short Name}
\item{ISTEST}{Immunogenicity Test or Examination Name}
\item{ISBDAGNT}{Binding Agent}
\item{ISCAT}{Category for Immunogenicity Test}
\item{ISORRES}{Results or Findings in Original Units}
\item{ISORRESU}{Original Units}
\item{ISSTRESC}{Character Result/Finding in Std Format}
\item{ISSTRESN}{Numeric Results/Findings in Std. Units}
\item{ISSTRESU}{Standard Units}
\item{ISSTAT}{Completion Status}
\item{ISREASND}{Reason Not Done}
\item{ISNAM}{Vendor Name}
\item{ISSPEC}{Specimen Type}
\item{ISBLFL}{Baseline Flag}
\item{ISLLOQ}{Lower Limit of Quantitation}
\item{VISITNUM}{Visit Number}
\item{VISIT}{Visit Name}
\item{VISITDY}{Planned Study Day of Visit}
\item{EPOCH}{Epoch}
\item{ISDTC}{Date/Time of Collection}
\item{ISDY}{Study Day of Visit/Collection/Exam}
\item{ISTPT}{Planned Time Point Name}
\item{ISTPTNUM}{Planned Time Point Number}
}
}
\source{
Generated dataset
}
\usage{
is_ada
}
\description{
A SDTM IS dataset containing relevant antidrug antibody assessment studies
}
\details{
Immunogenicity Specimen Assessments for ADA

A SDTM IS dataset containing relevant antidrug antibody assessment studies

Contains a set of 2 unique Test Short Names and Test Names: \tabular{ll}{
\strong{ISTESTCD} \tab \strong{ISTEST} \cr
ADA_BAB \tab Binding Antidrug Antibody             \cr
ADA_NAB \tab Neutralizing Binding Antidrug Antibody
}
}
\author{
Kristin Dahnert
}
\keyword{dataset}
