% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/is_vaccine.R
\docType{data}
\name{is_vaccine}
\alias{is_vaccine}
\title{Immunogenicity Specimen Assessments for Vaccine}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16 rows and 24 columns.

A data frame with 24 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{ISSEQ}{Sequence Number}
\item{ISTESTCD}{Immunogenicity Test/Exam Short Name}
\item{ISTEST}{Immunogenicity Test or Exam Name}
\item{ISCAT}{Category for Immunogenicity Test}
\item{ISORRES}{Result or Finding in Original Units}
\item{ISORRESU}{Original Units}
\item{ISSTRESC}{Character Result/Finding in Std Format}
\item{ISSTRESN}{Numeric Result/Finding in Standard Units}
\item{ISSTRESU}{Standard Units}
\item{ISSTAT}{Completion Status}
\item{ISREASND}{Reason Not Done}
\item{ISNAM}{Vendor Name}
\item{ISSPEC}{Specimen Type}
\item{ISMETHOD}{Method of Test or Examination}
\item{ISBLFL}{Baseline Flag}
\item{ISLLOQ}{Lower Limit of Quantitation}
\item{VISITNUM}{Visit Number}
\item{EPOCH}{Epoch}
\item{ISDTC}{Date/Time of Collection}
\item{ISDY}{Study Day of Collection}
\item{ISULOQ}{Upper Limit of Quantitation}
}
}
\source{
Constructed by \code{{admiralvaccine}} developers

Constructed by {admiralvaccine} developers
}
\usage{
is_vaccine

is_vaccine
}
\description{
An example SDTM IS for vaccine studies

Immunogenicity Specimen Assessments for Vaccine
}
\details{
An example SDTM IS for vaccine studies
}
\keyword{datasets}
