% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_metabolic.R
\docType{data}
\name{lb_metabolic}
\alias{lb_metabolic}
\title{Laboratory Measurements Dataset-metabolic}
\format{
A data frame with 23 columns:
\describe{
\item{STUDYID}{undocumented field}
\item{DOMAIN}{undocumented field}
\item{USUBJID}{undocumented field}
\item{LBSEQ}{undocumented field}
\item{LBTESTCD}{undocumented field}
\item{LBTEST}{undocumented field}
\item{LBCAT}{undocumented field}
\item{LBORRES}{undocumented field}
\item{LBORRESU}{undocumented field}
\item{LBORNRLO}{undocumented field}
\item{LBORNRHI}{undocumented field}
\item{LBSTRESC}{undocumented field}
\item{LBSTRESN}{undocumented field}
\item{LBSTRESU}{undocumented field}
\item{LBSTNRLO}{undocumented field}
\item{LBSTNRHI}{undocumented field}
\item{LBNRIND}{undocumented field}
\item{LBBLFL}{undocumented field}
\item{VISITNUM}{undocumented field}
\item{VISIT}{undocumented field}
\item{VISITDY}{undocumented field}
\item{LBDTC}{undocumented field}
\item{LBDY}{undocumented field}
}
}
\source{
Constructed by {admiralmetabolic} developers
}
\usage{
lb_metabolic
}
\description{
A SDTM LB dataset containing relevant laboratory measurements for metabolic studies
}
\details{
Laboratory Measurements Dataset-metabolic

A SDTM LB dataset containing relevant laboratory measurements for metabolic studies
}
\keyword{dataset}
