% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_neuro.R
\docType{data}
\name{lb_neuro}
\alias{lb_neuro}
\title{Laboratory Test Dataset-neuro}
\format{
A data frame with 23 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{LBSEQ}{Sequence Number}
\item{LBTESTCD}{Lab Test or Examination Short Name}
\item{LBTEST}{Lab Test or Examination Name}
\item{LBCAT}{Category for Lab Test}
\item{LBORRES}{Result or Finding in Original Units}
\item{LBORRESU}{Original Units}
\item{LBORNRLO}{Reference Range Lower Limit in Orig Unit}
\item{LBORNRHI}{Reference Range Upper Limit in Orig Unit}
\item{LBSTRESC}{Character Result/Finding in Std Format}
\item{LBSTRESN}{Numeric Result/Finding in Standard Units}
\item{LBSTRESU}{Standard Units}
\item{LBSTNRLO}{Reference Range Lower Limit-Std Units}
\item{LBSTNRHI}{Reference Range Upper Limit-Std Units}
\item{LBNRIND}{Reference Range Indicator}
\item{LBBLFL}{Baseline Flag}
\item{VISITNUM}{Visit Number}
\item{VISIT}{Visit Name}
\item{VISITDY}{Planned Study Day of Visit}
\item{LBDTC}{Date/Time of Specimen Collection}
\item{LBDY}{Study Day of Specimen Collection}
}
}
\source{
Constructed by {admiralneuro} developers
}
\usage{
lb_neuro
}
\description{
A SDTM LB domain dataset containing laboratory test results for Alzheimer's Disease studies
}
\details{
Laboratory Test Dataset-neuro

A SDTM LB domain dataset containing laboratory test results for Alzheimer's Disease studies

Contains a set of 4 unique Test Short Names and Test Names: \tabular{ll}{
\strong{LBTESTCD} \tab \strong{LBTEST} \cr
ABETA42  \tab Lumipulse G β-Amyloid 1-42-N Plasma             \cr
ASYNASAA \tab Alpha Synuclein Seed Amplification Assay (CSF)  \cr
PTARATIO \tab Lumipulse G pTau 217/β-Amyloid 1-42 Plasma Ratio\cr
PTAU217  \tab Lumipulse G pTau 217 Plasma
}
}
\keyword{dataset}
\keyword{neurology}
