% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_onco_pcwg3.R
\docType{data}
\name{lb_onco_pcwg3}
\alias{lb_onco_pcwg3}
\title{Laboratory Measurements (PSA) for Oncology}
\format{
A data frame with 18 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{LBSEQ}{Sequence Number}
\item{LBTESTCD}{Lab Test or Examination Short Name}
\item{LBTEST}{Lab Test or Examination Name}
\item{LBCAT}{Category for Lab Test}
\item{LBORRES}{Result or Finding in Original Units}
\item{LBORRESU}{Original Units}
\item{LBSTRESC}{Character Result/Finding in Std Format}
\item{LBSTRESN}{Numeric Result/Finding in Standard Units}
\item{LBSTRESU}{Standard Units}
\item{LBBLFL}{Baseline Flag}
\item{VISITNUM}{Visit Number}
\item{VISIT}{Visit Name}
\item{VISITDY}{Planned Study Day of Visit}
\item{LBDTC}{Date/Time of Specimen Collection}
\item{LBDY}{Study Day of Specimen Collection}
}
}
\source{
Generated dataset
}
\usage{
lb_onco_pcwg3
}
\description{
A SDTM LB dataset for prostate cancer studies intended for examples of ADaM dataset creation
}
\details{
Laboratory Measurements (PSA) for Oncology

A SDTM LB dataset for prostate cancer studies intended for examples of ADaM dataset creation

Contains a set of 1 unique Test Short Name and Test Name: \tabular{ll}{
\strong{LBTESTCD} \tab \strong{LBTEST} \cr
PSA \tab Prostate Specific Antigen
}
}
\author{
Tomoyuki Namai
}
\keyword{dataset}
\keyword{oncology}
