% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb.R
\docType{data}
\name{mb}
\alias{mb}
\title{Microbiology Specimen}
\format{
A data frame with 21 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{MBSEQ}{Sequence Number}
\item{MBGRPID}{Group ID}
\item{MBREFID}{Reference ID}
\item{MBLNKGRP}{Link Group ID}
\item{MBTESTCD}{Microbiology Test or Finding Short Name}
\item{MBTEST}{Microbiology Test or Finding Name}
\item{MBTSTDTL}{Measurement, Test or Examination Detail}
\item{MBORRES}{Original Result}
\item{MBORRESU}{Original Units}
\item{MBRSLSCL}{Result Scale}
\item{MBSTRESC}{Result or Finding in Standard Format}
\item{MBSTRESN}{Numeric Result/Finding in Standard Units}
\item{MBSTRESU}{Standard Units}
\item{MBSPEC}{Specimen Material Type}
\item{MBLOC}{Specimen Collection Location}
\item{MBMETHOD}{Method of Test or Examination}
\item{VISITNUM}{Visit Number}
\item{MBDTC}{Date/Time of Collection}
}
}
\source{
\href{https://github.com/pharmaverse/pharmaversesdtm/blob/main/data/mb.rda}{Access the source of the Microbiology Specimen dataset.}
}
\usage{
mb
}
\description{
A synthetic SDTM MB domain representing microbiology findings and linkage to MS domain (ms)
}
\details{
Microbiology Specimen

A synthetic SDTM MB domain representing microbiology findings and linkage to MS domain (ms)

Contains a set of 6 unique Test Short Names and Test Names: \tabular{ll}{
\strong{MBTESTCD} \tab \strong{MBTEST} \cr
GMNCOC   \tab Gram Negative Cocci               \cr
GNROD    \tab Gram Negative Rods                \cr
GPRCOC   \tab Gram Positive Cocci               \cr
MCCOLCNT \tab Colony Count                      \cr
MCORGIDN \tab Microbial Organism Identification \cr
MTBCMPLX \tab Mycobacterium tuberculosis complex
}
}
\keyword{dataset}
\keyword{microbiology}
