% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh.R
\docType{data}
\name{mh}
\alias{mh}
\title{Medical History}
\format{
A data frame with 28 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{MHSEQ}{Sequence Number}
\item{MHSPID}{Sponsor-Defined Identifier}
\item{MHTERM}{Reported Term for the Medical History}
\item{MHLLT}{Lowest Level Term}
\item{MHDECOD}{Dictionary-Derived Term}
\item{MHHLT}{High Level Term}
\item{MHHLGT}{High Level Group Term}
\item{MHCAT}{Category for Medical History}
\item{MHBODSYS}{Body System or Organ Class}
\item{MHSEV}{Severity/Intensity}
\item{VISITNUM}{Visit Number}
\item{VISIT}{Visit Name}
\item{VISITDY}{Planned Study Day of Visit}
\item{MHDTC}{Date/Time of History Collection}
\item{MHSTDTC}{Start Date/Time of Medical History Event}
\item{MHDY}{Study Day of History Collection}
\item{MHENDTC}{End Date/Time of Medical History Event}
\item{MHPRESP}{Medical History Event Pre-Specified}
\item{MHOCCUR}{Medical History Occurrence}
\item{MHSTRTPT}{Start Relative to Reference Time Point}
\item{MHENRTPT}{End Relative to Reference Time Point}
\item{MHSTTPT}{Start Reference Time Point}
\item{MHENTPT}{End Reference Time Point}
\item{MHENRF}{End Relative to Reference Period}
\item{MHSTAT}{Completion Status}
}
}
\source{
\href{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/mh.xpt?raw=true}{Access the source of the Medical History dataset.}
}
\usage{
mh
}
\description{
An updated SDTM MH dataset that uses data from the CDISC pilot project
}
\details{
Medical History

An updated SDTM MH dataset that uses data from the CDISC pilot project
}
\author{
Annie Yang
}
\keyword{dataset}
