% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms.R
\docType{data}
\name{ms}
\alias{ms}
\title{Microbiology Susceptibility}
\format{
A data frame with 23 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{MSSEQ}{Sequence Number}
\item{MSREFID}{Reference ID}
\item{NHOID}{Non-host Organism ID}
\item{MSGRPID}{Group ID}
\item{MSLNKID}{Link ID}
\item{MSTESTCD}{Short Name of Assessment}
\item{MSTEST}{Name of Assessment}
\item{MSAGENT}{Agent Name}
\item{MSCONC}{Agent Concentration}
\item{MSCONCU}{Agent Concentration Units}
\item{MSORRES}{Result or Finding in Original Units}
\item{MSORRESU}{Original Units}
\item{MSSTRESC}{Result or Finding in Standard Format}
\item{MSSTRESN}{Numeric Result/Finding in Standard Units}
\item{MSSTRESU}{Standard Units}
\item{MSSPEC}{Specimen Material Type}
\item{MSLOC}{Location Used for the Measurement}
\item{MSMETHOD}{Method of Test or Examination}
\item{VISITNUM}{Visit Number}
\item{MSDTC}{Date/Time of Collection}
}
}
\source{
\href{https://github.com/pharmaverse/pharmaversesdtm/blob/main/data/ms.rda}{Access the source of the Microbiology Susceptibility dataset.}
}
\usage{
ms
}
\description{
A synthetic SDTM MS domain with susceptibility results and linkage to MB domain (mb)
}
\details{
Microbiology Susceptibility

A synthetic SDTM MS domain with susceptibility results and linkage to MB domain (mb)

Contains a set of 3 unique Test Short Names and Test Names: \tabular{ll}{
\strong{MSTESTCD} \tab \strong{MSTEST} \cr
DIAZOINH \tab Diameter of the Zone of Inhibition\cr
MIC      \tab Minimum Inhibitory Concentration  \cr
MICROSUS \tab Microbial Susceptibility
}
}
\keyword{dataset}
\keyword{microbiology}
