% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nv_neuro.R
\docType{data}
\name{nv_neuro}
\alias{nv_neuro}
\title{Neurological Assessment Dataset-neuro}
\format{
A data frame with 21 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{NVSEQ}{Sequence Number}
\item{NVLNKID}{Link ID}
\item{NVTESTCD}{Short Name of Nervous System Test}
\item{NVTEST}{Name of Nervous System Test}
\item{NVCAT}{Category for Nervous System Test}
\item{NVLOC}{Location Used for the Measurement}
\item{NVMETHOD}{Method of Test or Examination}
\item{NVNAM}{Vendor Name}
\item{NVORRES}{Result or Finding in Original Units}
\item{NVORRESU}{Original Units}
\item{NVSTRESC}{Character Result/Finding in Std Format}
\item{NVSTRESN}{Numeric Result/Finding in Standard Units}
\item{NVSTRESU}{Standard Units}
\item{VISITNUM}{Visit Number}
\item{VISIT}{Visit Name}
\item{NVDTC}{Date/Time of Collection}
\item{NVDY}{Study Day of Collection}
\item{NVLOBXFL}{Last Observation Before Exposure Flag}
}
}
\source{
Constructed by admiralneuro developers
}
\usage{
nv_neuro
}
\description{
A SDTM NV dataset containing neurological assessments or test results
}
\details{
Neurological Assessment Dataset-neuro

A SDTM NV dataset containing neurological assessments or test results

Contains a set of 2 unique Test Short Names and Test Names: \tabular{ll}{
\strong{NVTESTCD} \tab \strong{NVTEST} \cr
SUVR \tab Standardized Uptake Value Ratio  \cr
VR   \tab Qualitative Visual Classification
}
}
\keyword{dataset}
