% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.R
\docType{data}
\name{pp}
\alias{pp}
\title{Pharmacokinetic Parameters}
\format{
A data frame with 14 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{PPSEQ}{Sequence Number}
\item{PPTESTCD}{Parameter Short Name}
\item{PPTEST}{Parameter Name}
\item{PPCAT}{Parameter Category}
\item{PPORRES}{Result or Finding in Original Units}
\item{PPORRESU}{Original Units}
\item{PPSTRESC}{Character Result/Finding in Std Format}
\item{PPSTRESN}{Numeric Result/Finding in Standard Units}
\item{PPSTRESU}{Standard Units}
\item{PPSPEC}{Specimen Material Type}
\item{PPRFDTC}{Date/Time of Reference Point}
}
}
\source{
\href{https://github.com/pharmaverse/pharmaversesdtm/blob/main/data/pp.rda}{Access the source of the Pharmacokinetic Parameters dataset.}
}
\usage{
pp
}
\description{
A SDTM PP dataset simulated by Antonio Rodriguez Contesti
}
\details{
Pharmacokinetic Parameters

A SDTM PP dataset simulated by Antonio Rodriguez Contesti

Contains a list of 10 unique test codes and names:
\describe{
\item{AUCALL}{AUC All}
\item{AUCLST}{AUC to Last Nonzero Conc}
\item{CLST}{Last Nonzero Conc}
\item{CMAX}{Max Conc}
\item{LAMZ}{Lambda z}
\item{LAMZHL}{Half-Life Lambda z}
\item{LAMZNPT}{Number of Points for Lambda z}
\item{RCAMINT}{Ae}
\item{RENALCL}{CLR}
\item{TMAX}{Time of CMAX}
}
}
\author{
Antonio Rodriguez Contesti
}
\keyword{dataset}
