% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qs_metabolic.R
\docType{data}
\name{qs_metabolic}
\alias{qs_metabolic}
\title{Questionnaire Dataset-metabolic}
\format{
A data frame with 18 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{USUBJID}{Unique Subject Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{VISIT}{Visit Name}
\item{VISITNUM}{Visit Number}
\item{VISITDY}{Planned Study Day of Visit}
\item{QSBLFL}{Baseline Flag}
\item{QSDTC}{Date/Time of Finding}
\item{QSDY}{Study Day of Finding}
\item{QSCAT}{Category for Questionnaire}
\item{QSTEST}{Questionnaire Test Name}
\item{QSTESTCD}{Questionnaire Test Short Name}
\item{QSORRES}{Result or Finding in Original Units}
\item{QSORRESU}{Original Units}
\item{QSSTRESC}{Character Result/Finding in Std Format}
\item{QSSTRESN}{Numeric Result/Finding in Standard Units}
\item{QSSTRESU}{Standard Units}
\item{QSSEQ}{Sequence Number}
}
}
\source{
Constructed by {admiralmetabolic} developers
}
\usage{
qs_metabolic
}
\description{
A SDTM QS dataset containing COEQ data (Control of Eating Questionnaire) for metabolic studies. Note that University of Leeds are the copyright holders of the Control of Eating Questionnaire (CoEQ) and the test data included within {pharmaversesdtm} as well as the ADCOEQ code are for not-for-profit use only within {admiralmetabolic} and pharmaverse-related examples/documentation. Any persons or companies wanting to use the CoEQ should request a license to do so from the following \href{https://licensing.leeds.ac.uk/product/control-of-eating-questionnaire-coeq}{link}.
}
\details{
Questionnaire Dataset-metabolic

A SDTM QS dataset containing COEQ data (Control of Eating Questionnaire) for metabolic studies. Note that University of Leeds are the copyright holders of the Control of Eating Questionnaire (CoEQ) and the test data included within {pharmaversesdtm} as well as the ADCOEQ code are for not-for-profit use only within {admiralmetabolic} and pharmaverse-related examples/documentation. Any persons or companies wanting to use the CoEQ should request a license to do so from the following \href{https://licensing.leeds.ac.uk/product/control-of-eating-questionnaire-coeq}{link}.

Contains a list of 21 unique Test Short Names (QSTESTCD) and Test Names (QSTEST):
\describe{
\item{COEQ01}{How hungry have you felt?}
\item{COEQ02}{How full have you felt?}
\item{COEQ03}{How strong was your desire to eat sweet foods?}
\item{COEQ04}{How strong was your desire to eat savoury foods?}
\item{COEQ05}{How happy have you felt?}
\item{COEQ06}{How anxious have you felt?}
\item{COEQ07}{How alert have you felt?}
\item{COEQ08}{How contented have you felt?}
\item{COEQ09}{During the last 7 days how often have you had food cravings?}
\item{COEQ10}{How strong have any food cravings been?}
\item{COEQ11}{How difficult has it been to resist any food cravings?}
\item{COEQ12}{How often have you eaten in response to food cravings?}
\item{COEQ13}{Chocolate or chocolate flavoured foods}
\item{COEQ14}{Other sweet foods (cakes, pastries, biscuits, etc)}
\item{COEQ15}{Fruit or fruit juice}
\item{COEQ16}{Dairy foods (cheese, yoghurts, milk, etc)}
\item{COEQ17}{Starchy foods (bread, rice, pasta, etc)}
\item{COEQ18}{Savoury foods (french fries, crisps, burgers, pizza, etc)}
\item{COEQ19}{Generally, how difficult has it been to control your eating?}
\item{COEQ20}{Which one food makes it most difficult for you to control eating?}
\item{COEQ21}{How difficult has it been to resist eating this food during the last 7 days?}
}
}
\keyword{dataset}
