% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qs_ophtha.R
\docType{data}
\name{qs_ophtha}
\alias{qs_ophtha}
\title{Questionnaire for Ophthalmology}
\format{
A data frame with 20 columns:
\describe{
\item{STUDYID}{undocumented field}
\item{DOMAIN}{Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{QSSEQ}{Sequence Number}
\item{QSTESTCD}{Question Short Name}
\item{QSTEST}{Question Name}
\item{QSCAT}{Category of Question}
\item{QSSCAT}{Subcategory for Question}
\item{QSORRES}{Finding in Original Units}
\item{QSORRESU}{Original Units}
\item{QSSTRESC}{Character Result/Finding in Std Format}
\item{QSSTRESN}{Numeric Finding in Standard Units}
\item{QSSTRESU}{Standard Units}
\item{QSBLFL}{Baseline Flag}
\item{QSDRVFL}{Derived Flag}
\item{VISITNUM}{Visit Number}
\item{VISIT}{Visit Name}
\item{VISITDY}{Planned Study Day of Visit}
\item{QSDTC}{Date/Time of Finding}
\item{QSDY}{Study Day of Finding}
}
}
\source{
Constructed using qs from the {pharmaversesdtm} package
}
\usage{
qs_ophtha
}
\description{
An example Questionnaires SDTM dataset with ophthalmology-specific questionnaire of NEI VFQ-25
}
\details{
Questionnaire for Ophthalmology

An example Questionnaires SDTM dataset with ophthalmology-specific questionnaire of NEI VFQ-25
}
\keyword{dataset}
