% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_ds}
\alias{raw_ds}
\title{Disposition Dataset-raw}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 596 rows and 13 columns.
}
\source{
\url{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/ds.xpt?raw=true} # nolint
}
\usage{
raw_ds
}
\description{
A SDTM DS dataset from the CDISC pilot project
}
\keyword{datasets}
