% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_mh}
\alias{raw_mh}
\title{Medical History Dataset-raw}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1818 rows and 19 columns.
}
\source{
\url{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/mh.xpt?raw=true} # nolint
}
\usage{
raw_mh
}
\description{
A SDTM MH dataset from the CDISC pilot project
}
\keyword{datasets}
