% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_qs}
\alias{raw_qs}
\title{Questionnaire Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 121749 rows and 20 columns.
}
\source{
\url{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/qs.xpt?raw=true} # nolint
}
\usage{
raw_qs
}
\description{
A SDTM QS dataset from the CDISC pilot project
}
\keyword{datasets}
