% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/rs_onco.R
\docType{data}
\name{rs_onco}
\alias{rs_onco}
\title{Disease Response for Oncology}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5808 rows and 19 columns.

A data frame with 19 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{RSSEQ}{Sequence Number}
\item{RSLNKGRP}{Link Group}
\item{RSTESTCD}{Response Assessment Short Name}
\item{RSTEST}{Response Assessment Name}
\item{RSCAT}{Category for Response Assessment}
\item{RSORRES}{Response Assessment Original Result}
\item{RSSTRESC}{Response Assessment Result in Std Format}
\item{RSSTAT}{Completion Status}
\item{RSREASND}{Reason Response Assessment Not Performed}
\item{RSEVAL}{Evaluator}
\item{RSEVALID}{Evaluator Identifier}
\item{RSACPTFL}{Accepted Record Flag}
\item{VISITNUM}{Visit Number}
\item{VISIT}{Visit Name}
\item{RSDTC}{Date/Time of Response Assessment}
\item{RSDY}{Study Day of Response Assessment}
}
}
\source{
https://github.com/pharmaverse/pharmaversesdtm/blob/main/data/tr_onco_recist.rda
}
\usage{
rs_onco

rs_onco
}
\description{
A SDTM RS dataset simulated by Gopi Vegesna

Disease Response for Oncology
}
\details{
A SDTM RS dataset simulated by Gopi Vegesna
}
\author{
Gopi Vegesna
}
\keyword{datasets}
