% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rs_onco_ca125}
\alias{rs_onco_ca125}
\title{Disease Response (GCIG)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 66 rows and 13 columns.
}
\usage{
rs_onco_ca125
}
\description{
A SDTM RS dataset using GCIG criteria. The dataset contains just a few patients.
It is intended for vignettes and examples of ADaM dataset creation.
}
\author{
Vinh Nguyen
}
\keyword{datasets}
