% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/rs_onco_imwg.R
\docType{data}
\name{rs_onco_imwg}
\alias{rs_onco_imwg}
\title{Disease Response (IMWG)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 65 rows and 17 columns.

A data frame with 17 columns:
\describe{
\item{STUDYID}{undocumented field}
\item{DOMAIN}{undocumented field}
\item{USUBJID}{undocumented field}
\item{RSSEQ}{undocumented field}
\item{RSLNKGRP}{undocumented field}
\item{RSTESTCD}{undocumented field}
\item{RSTEST}{undocumented field}
\item{RSCAT}{undocumented field}
\item{RSORRES}{undocumented field}
\item{RSSTRESC}{undocumented field}
\item{RSSTAT}{undocumented field}
\item{RSREASND}{undocumented field}
\item{RSEVAL}{undocumented field}
\item{VISITNUM}{undocumented field}
\item{VISIT}{undocumented field}
\item{RSDTC}{undocumented field}
\item{RSDY}{undocumented field}
}
}
\source{
Derived from tr_onco_recist
}
\usage{
rs_onco_imwg

rs_onco_imwg
}
\description{
A SDTM RS dataset using IMWG criteria. The dataset contains just a few patients.
It is intended for vignettes and examples of ADaM dataset creation.

Disease Response (IMWG)
}
\details{
A SDTM RS dataset using IMWG criteria intended for examples of ADaM dataset creation
}
\author{
Vinh Nguyen
}
\keyword{datasets}
