% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_onco_lymphoma.R
\docType{data}
\name{rs_onco_lymphoma}
\alias{rs_onco_lymphoma}
\title{Disease Response (Lymphoma Lugano 2014)}
\format{
A data frame with 17 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{RSSEQ}{Sequence Number}
\item{RSTESTCD}{Response Assessment Short Name}
\item{RSTEST}{Response Assessment Name}
\item{RSCAT}{Category for Response Assessment}
\item{RSMETHOD}{undocumented field}
\item{RSSCAT}{undocumented field}
\item{RSORRES}{Response Assessment Original Result}
\item{RSSTRESC}{Response Assessment Result in Std Format}
\item{RSSTAT}{Completion Status}
\item{RSEVAL}{Evaluator}
\item{VISITNUM}{Visit Number}
\item{VISIT}{Visit Name}
\item{RSDTC}{Date/Time of Response Assessment}
\item{RSDY}{Study Day of Response Assessment}
}
}
\source{
Generated dataset.
}
\usage{
rs_onco_lymphoma
}
\description{
A SDTM RS domain for lymphoma studies using Lugano 2014 response criteria intended for examples of ADaM dataset creation
}
\details{
Disease Response (Lymphoma Lugano 2014)

A SDTM RS domain for lymphoma studies using Lugano 2014 response criteria intended for examples of ADaM dataset creation

Contains a set of 1 unique Test Short Name and Test Name: \tabular{ll}{
\strong{RSTESTCD} \tab \strong{RSTEST} \cr
OVRLRESP \tab Overall Response
}
}
\author{
Kiran Peddamudium
}
\keyword{dataset}
\keyword{oncology}
