% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_onco_pcwg3.R
\docType{data}
\name{rs_onco_pcwg3}
\alias{rs_onco_pcwg3}
\title{Disease Response (PCWG3) for Oncology}
\format{
A data frame with 14 columns:
\describe{
\item{STUDYID}{undocumented field}
\item{DOMAIN}{undocumented field}
\item{USUBJID}{undocumented field}
\item{RSSEQ}{undocumented field}
\item{RSTESTCD}{undocumented field}
\item{RSTEST}{undocumented field}
\item{RSCAT}{undocumented field}
\item{RSORRES}{undocumented field}
\item{RSSTRESC}{undocumented field}
\item{RSEVAL}{undocumented field}
\item{VISITNUM}{Visit Number}
\item{VISIT}{undocumented field}
\item{RSDTC}{undocumented field}
\item{RSDY}{Study Day of Specimen Collection}
}
}
\source{
Generated dataset
}
\usage{
rs_onco_pcwg3
}
\description{
A SDTM RS dataset for oncology studies using PCWG3 criteria intended for examples of ADaM dataset creation
}
\details{
Disease Response (PCWG3) for Oncology

A SDTM RS dataset for oncology studies using PCWG3 criteria intended for examples of ADaM dataset creation
}
\author{
Tomoyuki Namai
}
\keyword{dataset}
