% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/rs_onco_recist.R
\docType{data}
\name{rs_onco_recist}
\alias{rs_onco_recist}
\title{Disease Response (RECIST 1.1) for Oncology}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 66 rows and 14 columns.

A data frame with 14 columns:
\describe{
\item{DOMAIN}{undocumented field}
\item{STUDYID}{undocumented field}
\item{USUBJID}{undocumented field}
\item{VISITNUM}{undocumented field}
\item{VISIT}{undocumented field}
\item{RSTESTCD}{undocumented field}
\item{RSTEST}{undocumented field}
\item{RSORRES}{undocumented field}
\item{RSSTRESC}{undocumented field}
\item{RSEVAL}{undocumented field}
\item{RSEVALID}{undocumented field}
\item{RSACPTFL}{undocumented field}
\item{RSDTC}{undocumented field}
\item{RSSEQ}{undocumented field}
}
}
\source{
The dataset is derived from \code{tr_onco_recist}.

Generated dataset.
}
\usage{
rs_onco_recist

rs_onco_recist
}
\description{
A SDTM RS dataset using RECIST 1.1. The dataset contains just a few patients.
It is intended for vignettes and examples of ADaM dataset creation.

Disease Response (RECIST 1.1) for Oncology
}
\details{
A SDTM RS dataset using RECIST 1.1 intended for examples of ADaM dataset creation
}
\author{
Stefan Bundfuss
}
\keyword{datasets}
