% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_ophtha.R
\docType{data}
\name{sc_ophtha}
\alias{sc_ophtha}
\title{Subject Characteristic for Ophthalmology}
\format{
A data frame with 12 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{SCSEQ}{Sequence Number}
\item{SCTESTCD}{Subject Characteristic Short Name}
\item{SCTEST}{Subject Characteristic}
\item{SCCAT}{Category for Subject Characteristic}
\item{SCORRES}{Result or Finding in Original Units}
\item{SCSTRESC}{Character Result/Finding in Std Format}
\item{EPOCH}{Epoch}
\item{SCDTC}{Date/Time of Collection}
\item{SCDY}{Study Day of Examination}
}
}
\source{
Generated dataset.
}
\usage{
sc_ophtha
}
\description{
A SDTM SC dataset simulated by Ophthalmology team
}
\details{
Subject Characteristic for Ophthalmology

A SDTM SC dataset simulated by Ophthalmology team
}
\author{
Gordon Miller
}
\keyword{dataset}
