% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdg_db.R
\docType{data}
\name{sdg_db}
\alias{sdg_db}
\title{sdg_db}
\format{
A data frame with 5 columns:
\describe{
\item{termchar}{undocumented field}
\item{sdg_name}{undocumented field}
\item{sdg_id}{undocumented field}
\item{termvar}{undocumented field}
\item{version}{undocumented field}
}
}
\source{
https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/sv.xpt?raw=true # nolint
}
\usage{
sdg_db
}
\description{
SDG
}
\details{
An example SDG dataset
}
\keyword{datasets}
