% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/smq_db.R
\docType{data}
\name{smq_db}
\alias{smq_db}
\title{Standardized MedDRA Queries}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 44 rows and 6 columns.

A data frame with 6 columns:
\describe{
\item{termchar}{undocumented field}
\item{scope}{undocumented field}
\item{smq_name}{undocumented field}
\item{smq_id}{undocumented field}
\item{version}{undocumented field}
\item{termvar}{undocumented field}
}
}
\source{
Generated dataset.
}
\usage{
smq_db

smq_db
}
\description{
An example SMQ dataset

Standardized MedDRA Queries
}
\details{
An example SMQ dataset
}
\keyword{datasets}
