% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/suppae.R
\docType{data}
\name{suppae}
\alias{suppae}
\title{Supplemental Adverse Events}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1191 rows and 10 columns.

A data frame with 10 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{RDOMAIN}{Related Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{IDVAR}{Identifying Variable}
\item{IDVARVAL}{Identifying Variable Value}
\item{QNAM}{Qualifier Variable Name}
\item{QLABEL}{Qualifier Variable Label}
\item{QVAL}{Data Value}
\item{QORIG}{Origin}
\item{QEVAL}{Evaluator}
}
}
\source{
\url{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/suppae.xpt?raw=true}  # nolint

https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/suppds.xpt?raw=true
}
\usage{
suppae

suppae
}
\description{
A SDTM SUPPAE dataset from the CDISC pilot project

Supplemental Adverse Events
}
\details{
A SDTM SUPPAE dataset from the CDISC pilot project
}
\keyword{datasets}
