% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/suppce_vaccine.R
\docType{data}
\name{suppce_vaccine}
\alias{suppce_vaccine}
\title{Supplemental Qualifiers for Clinical Events for Vaccine}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 9 columns.

A data frame with 9 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{USUBJID}{Unique Subject Identifier}
\item{RDOMAIN}{Related Domain Abbreviation}
\item{IDVAR}{Identifying Variable}
\item{IDVARVAL}{Identifying Variable Value}
\item{QNAM}{Qualifier Variable Name}
\item{QVAL}{Data Value}
\item{QLABEL}{Qualifier Variable Label}
\item{QORIG}{Origin}
}
}
\source{
Constructed by \code{{admiralvaccine}} developers

Constructed by {admiralvaccine} developers
}
\usage{
suppce_vaccine

suppce_vaccine
}
\description{
An example SDTM SUPPCE for vaccine studies

Supplemental Qualifiers for Clinical Events for Vaccine
}
\details{
An example SDTM SUPPCE for vaccine studies
}
\keyword{datasets}
