% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/suppdm.R
\docType{data}
\name{suppdm}
\alias{suppdm}
\title{Supplemental Demography}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1197 rows and 10 columns.

A data frame with 10 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{RDOMAIN}{Related Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{IDVAR}{Identifying Variable}
\item{IDVARVAL}{Identifying Variable Value}
\item{QNAM}{Qualifier Variable Name}
\item{QLABEL}{Qualifier Variable Label}
\item{QVAL}{Data Value}
\item{QORIG}{Origin}
\item{QEVAL}{Evaluator}
}
}
\source{
\url{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/suppdm.xpt?raw=true}  # nolint

Generated dataset.
}
\usage{
suppdm

suppdm
}
\description{
A SDTM SUPPDM dataset from the CDISC pilot project

Supplemental Demography
}
\details{
A SDTM SUPPDM dataset from the CDISC pilot project
}
\keyword{datasets}
