% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/suppis_vaccine.R
\docType{data}
\name{suppis_vaccine}
\alias{suppis_vaccine}
\title{Supplemental Qualifiers for Immunogenicity Specimen Assessments for Vaccine}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16 rows and 10 columns.

A data frame with 10 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{RDOMAIN}{Related Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{IDVAR}{Identifying Variable}
\item{IDVARVAL}{Identifying Variable Value}
\item{QNAM}{Qualifier Variable Name}
\item{QLABEL}{Qualifier Variable Label}
\item{QVAL}{Data Value}
\item{QORIG}{Origin}
\item{QEVAL}{Evaluator}
}
}
\source{
Constructed by \code{{admiralvaccine}} developers

Constructed by {admiralpeds} developers
}
\usage{
suppis_vaccine

suppis_vaccine
}
\description{
An example SDTM SUPPIS dataset for vaccine studies

Supplemental Qualifiers for Immunogenicity Specimen Assessments for Vaccine
}
\details{
An example SDTM SUPPIS dataset for vaccine studies
}
\keyword{datasets}
