% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppnv_neuro.R
\docType{data}
\name{suppnv_neuro}
\alias{suppnv_neuro}
\title{Supplemental for Neurological Assessment Dataset-neuro}
\format{
A data frame with 10 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{RDOMAIN}{Related Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{IDVAR}{Identifying Variable}
\item{IDVARVAL}{Identifying Variable Value}
\item{QNAM}{Qualifier Variable Name}
\item{QLABEL}{Qualifier Variable Label}
\item{QVAL}{Data Value}
\item{QORIG}{Origin}
\item{QEVAL}{Evaluator}
}
}
\source{
Constructed by admiralneuro developers
}
\usage{
suppnv_neuro
}
\description{
A SDTM SUPPNV dataset containing additional information about neurological assessments or test results
}
\details{
Supplemental for Neurological Assessment Dataset-neuro

A SDTM SUPPNV dataset containing additional information about neurological assessments or test results
}
\keyword{dataset}
