% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supprs_onco_ca125.R
\docType{data}
\name{supprs_onco_ca125}
\alias{supprs_onco_ca125}
\title{No label available}
\format{
A data frame with 9 columns:
\describe{
\item{STUDYID}{undocumented field}
\item{RDOMAIN}{undocumented field}
\item{USUBJID}{undocumented field}
\item{IDVAR}{undocumented field}
\item{IDVARVAL}{undocumented field}
\item{QNAM}{undocumented field}
\item{QLABEL}{undocumented field}
\item{QVAL}{undocumented field}
\item{QORIG}{undocumented field}
}
}
\source{
No source available
}
\usage{
supprs_onco_ca125
}
\description{
dataset
}
\details{
No label available

No description available
}
\keyword{dataset}
