% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/supprs_onco_imwg.R
\docType{data}
\name{supprs_onco_imwg}
\alias{supprs_onco_imwg}
\title{Supplemental Qualifiers for RS_ONCO_IMWG}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 19 rows and 9 columns.

A data frame with 9 columns:
\describe{
\item{STUDYID}{undocumented field}
\item{RDOMAIN}{undocumented field}
\item{USUBJID}{undocumented field}
\item{IDVAR}{undocumented field}
\item{IDVARVAL}{undocumented field}
\item{QNAM}{undocumented field}
\item{QLABEL}{undocumented field}
\item{QVAL}{undocumented field}
\item{QORIG}{undocumented field}
}
}
\source{
https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/suppae.xpt?raw=true
}
\usage{
supprs_onco_imwg

supprs_onco_imwg
}
\description{
A SDTM supplemental RS dataset using IMWG criteria. It is intended to be used
together with \code{rs_onco_imwg}.

Supplemental Qualifiers for RS_ONCO_IMWG
}
\details{
A SDTM supplemental RS dataset using IMWG criteria intended to be used with rs_onco_imwg
}
\author{
Vinh Nguyen
}
\keyword{datasets}
