% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/supptr_onco.R
\docType{data}
\name{supptr_onco}
\alias{supptr_onco}
\title{Supplemental Tumor Results for Oncology}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 55995 rows and 9 columns.

A data frame with 9 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{RDOMAIN}{Related Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{IDVAR}{Identifying Variable}
\item{IDVARVAL}{Identifying Variable Value}
\item{QNAM}{Qualifier Variable Name}
\item{QLABEL}{Qualifier Variable Label}
\item{QVAL}{Data Value}
\item{QORIG}{Origin}
}
}
\source{
https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/ts.xpt?raw=true
}
\usage{
supptr_onco

supptr_onco
}
\description{
A SDTM SUPPTR dataset simulated by Gopi Vegesna

Supplemental Tumor Results for Oncology
}
\details{
A SDTM SUPPTR dataset simulated by Gopi Vegesna
}
\author{
Gopi Vegesna
}
\keyword{datasets}
