% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/sv.R
\docType{data}
\name{sv}
\alias{sv}
\title{Subject Visits}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3559 rows and 8 columns.

A data frame with 8 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{VISITNUM}{Visit Number}
\item{VISIT}{Visit Name}
\item{VISITDY}{Planned Study Day of Visit}
\item{SVSTDTC}{Start Date/Time of Visit}
\item{SVENDTC}{End Date/Time of Visit}
}
}
\source{
\url{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/sv.xpt?raw=true}  # nolint

Constructed by {admiralvaccine} developers
}
\usage{
sv

sv
}
\description{
A SDTM SV dataset from the CDISC pilot project. Duplicate observation for group
variable \code{USUBJID} and \code{VISIT} is corrected.

Subject Visits
}
\details{
A SDTM SV dataset from the CDISC pilot project with corrected duplicate observation
}
\keyword{datasets}
