% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_onco.R
\docType{data}
\name{tr_onco}
\alias{tr_onco}
\title{Tumor Results for Oncology}
\format{
A data frame with 24 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{TRSEQ}{Sequence Number}
\item{TRGRPID}{Group ID}
\item{TRLNKGRP}{Link Group}
\item{TRLNKID}{Link ID}
\item{TRTESTCD}{Tumor Assessment Short Name}
\item{TRTEST}{Tumor Assessment Test Name}
\item{TRORRES}{Result or Finding in Original Units}
\item{TRORRESU}{Original Units}
\item{TRSTRESC}{Character Result/Finding in Std Format}
\item{TRSTRESN}{Numeric Result/Finding in Standard Units}
\item{TRSTRESU}{Standard Units}
\item{TRSTAT}{Completion Status}
\item{TRREASND}{Reason Tumor Measurement Not Performed}
\item{TRMETHOD}{Method used to Identify the Tumor}
\item{TREVAL}{Evaluator}
\item{TREVALID}{Evaluator Identifier}
\item{TRACPTFL}{Accepted Record Flag}
\item{VISITNUM}{Visit Number}
\item{VISIT}{Visit Name}
\item{TRDTC}{Date/Time of Tumor Measurement}
\item{TRDY}{Study Day of Tumor Measurement}
}
}
\source{
Generated dataset.
}
\usage{
tr_onco
}
\description{
A SDTM TR dataset simulated by Gopi Vegesna
}
\details{
Tumor Results for Oncology

A SDTM TR dataset simulated by Gopi Vegesna
}
\author{
Gopi Vegesna
}
\keyword{dataset}
