% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/ts.R
\docType{data}
\name{ts}
\alias{ts}
\title{Trial Design}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 33 rows and 6 columns.

A data frame with 6 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{TSSEQ}{Sequence Number}
\item{TSPARMCD}{Trial Summary Parameter Short Name}
\item{TSPARM}{Trial Summary Parameter}
\item{TSVAL}{Parameter Value}
}
}
\source{
\url{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/ts.xpt?raw=true}  # nolint

https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/vs.xpt?raw=true
}
\usage{
ts

ts
}
\description{
A SDTM TS dataset from the CDISC pilot project

Trial Design
}
\details{
A SDTM TS dataset from the CDISC pilot project
}
\keyword{datasets}
