% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/tu_onco_recist.R
\docType{data}
\name{tu_onco_recist}
\alias{tu_onco_recist}
\title{Tumor Identification (RECIST 1.1) for Oncology}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 75 rows and 16 columns.

A data frame with 16 columns:
\describe{
\item{DOMAIN}{undocumented field}
\item{STUDYID}{undocumented field}
\item{USUBJID}{undocumented field}
\item{VISIT}{undocumented field}
\item{VISITNUM}{undocumented field}
\item{TULOC}{undocumented field}
\item{TUTESTCD}{undocumented field}
\item{TUTEST}{undocumented field}
\item{TUORRES}{undocumented field}
\item{TUSTRESC}{undocumented field}
\item{TUMETHOD}{undocumented field}
\item{TULNKID}{undocumented field}
\item{TUEVAL}{undocumented field}
\item{TUEVALID}{undocumented field}
\item{TUACPTFL}{undocumented field}
\item{TUSEQ}{undocumented field}
}
}
\source{
Generated dataset.
}
\usage{
tu_onco_recist

tu_onco_recist
}
\description{
A SDTM TU dataset using RECIST 1.1. The dataset contains just a few patients.
It is intended for vignettes and examples of ADaM dataset creation.

Tumor Identification (RECIST 1.1) for Oncology
}
\details{
A SDTM TU dataset using RECIST 1.1 intended for examples of ADaM dataset creation
}
\author{
Stefan Bundfuss
}
\keyword{datasets}
