% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vs}
\alias{vs}
\title{Vital Signs Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 29643 rows and 24 columns.
}
\source{
\url{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/vs.xpt?raw=true} # nolint
}
\usage{
vs
}
\description{
A SDTM VS dataset from the CDISC pilot project
}
\keyword{datasets}
