% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs_metabolic.R
\docType{data}
\name{vs_metabolic}
\alias{vs_metabolic}
\title{Vital signs Dataset-metabolic}
\format{
A data frame with 24 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{VSSEQ}{Sequence Number}
\item{VSTESTCD}{Vital Signs Test Short Name}
\item{VSTEST}{Vital Signs Test Name}
\item{VSPOS}{Vital Signs Position of Subject}
\item{VSORRES}{Result or Finding in Original Units}
\item{VSORRESU}{Original Units}
\item{VSSTRESC}{Character Result/Finding in Std Format}
\item{VSSTRESN}{Numeric Result/Finding in Standard Units}
\item{VSSTRESU}{Standard Units}
\item{VSSTAT}{Completion Status}
\item{VSLOC}{Location of Vital Signs Measurement}
\item{VSBLFL}{Baseline Flag}
\item{VISITNUM}{Visit Number}
\item{VISIT}{Visit Name}
\item{VISITDY}{Planned Study Day of Visit}
\item{VSDTC}{Date/Time of Measurements}
\item{VSDY}{Study Day of Vital Signs}
\item{VSTPT}{Planned Time Point Name}
\item{VSTPTNUM}{Planned Time Point Number}
\item{VSELTM}{Planned Elapsed Time from Time Point Ref}
\item{VSTPTREF}{Time Point Reference}
}
}
\source{
Constructed by {admiralmetabolic} developers
}
\usage{
vs_metabolic
}
\description{
A SDTM VS dataset for metabolic studies
}
\details{
Vital signs Dataset-metabolic

A SDTM VS dataset for metabolic studies

Contains a set of 9 unique Test Short Names and Test Names: \tabular{ll}{
\strong{VSTESTCD} \tab \strong{VSTEST} \cr
BMI    \tab Body Mass Index         \cr
DIABP  \tab Diastolic Blood Pressure\cr
HEIGHT \tab Height                  \cr
HIPCIR \tab Hip Circumference       \cr
PULSE  \tab Pulse Rate              \cr
SYSBP  \tab Systolic Blood Pressure \cr
TEMP   \tab Temperature             \cr
WEIGHT \tab Weight                  \cr
WSTCIR \tab Waist Circumference
}
}
\keyword{dataset}
