% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ce_vaccine}
\alias{ce_vaccine}
\title{Clinical Events for Vaccine}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 44 rows and 29 columns.
}
\source{
Constructed by \code{{admiralvaccine}} developers
}
\usage{
ce_vaccine
}
\description{
An example SDTM CE dataset for vaccine studies
}
\keyword{datasets}
