% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dm_vaccine}
\alias{dm_vaccine}
\title{Demographics for Vaccine}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2 rows and 28 columns.
}
\source{
Constructed by \code{{admiralvaccine}} developers
}
\usage{
dm_vaccine
}
\description{
An example SDTM DM dataset for vaccine studies
}
\keyword{datasets}
