% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eg.R
\docType{data}
\name{eg}
\alias{eg}
\title{Electrocardiogram}
\format{
A data frame with 23 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{EGSEQ}{Sequence Number}
\item{EGTESTCD}{Test Code}
\item{EGTEST}{Test Name}
\item{EGORRES}{Original Result}
\item{EGORRESU}{Original Units}
\item{EGSTRESC}{Standard Character Result}
\item{EGSTRESN}{Standard Numeric Result}
\item{EGSTRESU}{Standard Units}
\item{EGSTAT}{Completion Status}
\item{EGLOC}{Location of Vital Signs Measurement}
\item{EGBLFL}{Baseline Flag}
\item{VISITNUM}{Visit Number}
\item{VISIT}{Visit Name}
\item{VISITDY}{Planned Study Day of Visit}
\item{EGDTC}{Date/Time of Measurements}
\item{EGDY}{Study Day of Vital Signs}
\item{EGTPT}{Planned Time Point Number}
\item{EGTPTNUM}{Time Point Number}
\item{EGELTM}{Planned Elapsed Time from Time Point Ref}
\item{EGTPTREF}{Time Point Reference}
}
}
\source{
Generated dataset.
}
\usage{
eg
}
\description{
An example of standard SDTM EG dataset to be used in deriving ADEG dataset
}
\details{
Electrocardiogram

An example of standard SDTM EG dataset to be used in deriving ADEG dataset
}
\keyword{dataset}
