% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_ophtha}
\alias{ex_ophtha}
\title{Ophthalmology Exposure Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 591 rows and 19 columns.
}
\source{
Constructed using `ex` from the `{pharmaversesdtm}` package # nolint
}
\usage{
ex_ophtha
}
\description{
An example Exposure SDTM dataset with ophthalmology-specific variables such as `EXLOC` and `EXLAT`
}
\keyword{datasets}
