% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_vaccine}
\alias{ex_vaccine}
\title{Exposures Dataset for Vaccine Studies}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 21 columns.
}
\source{
\url{https://github.com/pharmaverse/admiralvaccine/blob/main/data/vx_ex.rda} # nolint
}
\usage{
ex_vaccine
}
\description{
An example SDTM EX dataset for vaccine studies
}
\keyword{datasets}
