% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{face_vaccine}
\alias{face_vaccine}
\title{Findings About Clinical Events Dataset for Vaccine Studies}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 307 rows and 30 columns.
}
\source{
\url{https://github.com/pharmaverse/admiralvaccine/blob/main/data/vx_face.rda} # nolint
}
\usage{
face_vaccine
}
\description{
An example SDTM FACE dataset for vaccine studies
}
\keyword{datasets}
