% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{is_vaccine}
\alias{is_vaccine}
\title{Immunogenicity Specimen Assessments Dataset for Vaccine Studies}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16 rows and 24 columns.
}
\source{
\url{https://github.com/pharmaverse/admiralvaccine/blob/main/data/vx_is.rda} # nolint
}
\usage{
is_vaccine
}
\description{
An example SDTM IS dataset for vaccine studies
}
\keyword{datasets}
