% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mh}
\alias{mh}
\title{Medical History Dataset-updated}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1818 rows and 28 columns.
}
\source{
\url{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/mh.xpt?raw=true} # nolint
}
\usage{
mh
}
\description{
An updated SDTM MH dataset that uses data from the CDISC pilot project
}
\author{
Annie Yang
}
\keyword{datasets}
