% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oe_ophtha}
\alias{oe_ophtha}
\title{Ophthalmic Examinations for Ophthalmology}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 15344 rows and 25 columns.
}
\usage{
oe_ophtha
}
\description{
A SDTM OE dataset simulated by Ophthalmology team
}
\author{
Gordon Miller
}
\keyword{datasets}
