% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oe_ophtha.R
\docType{data}
\name{oe_ophtha}
\alias{oe_ophtha}
\title{Ophthalmic Examinations for Ophthalmology}
\format{
A data frame with 25 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{OESEQ}{Sequence Number}
\item{OECAT}{Category for Ophthalmic Test or Exam}
\item{OESCAT}{Subcategory for Ophthalmic Test or Exam}
\item{OEDTC}{Date/Time of Collection}
\item{VISIT}{Visit Name}
\item{VISITNUM}{Visit Number}
\item{VISITDY}{Planned Study Day of Visit}
\item{OESTRESN}{Numeric Result/Finding in Standard Units}
\item{OESTRESC}{Character Result/Finding in Std Format}
\item{OEORRES}{Result or Finding in Original Units}
\item{OETEST}{Name of Ophthalmic Test or Examination}
\item{OETESTCD}{Short Name of Ophthalmic Test or Exam}
\item{OETSTDTL}{Ophthalmic Test or Exam Detail}
\item{OELAT}{Laterality}
\item{OELOC}{Location Used for the Measurement}
\item{OEDY}{Study Day of Visit/Collection/Exam}
\item{OEMETHOD}{Method of Test or Examination}
\item{OEORRESU}{Original Units}
\item{OESTRESU}{Standard Units}
\item{OESTAT}{Completion Status}
\item{OETPT}{Planned Time Point Name}
\item{OETPTNUM}{Planned Time Point Number}
}
}
\source{
Generated dataset.
}
\usage{
oe_ophtha
}
\description{
A SDTM OE dataset simulated by Ophthalmology team
}
\details{
Ophthalmic Examinations for Ophthalmology

A SDTM OE dataset simulated by Ophthalmology team

Contains a set of 4 unique Test Short Names and Test Names: \tabular{ll}{
\strong{OETESTCD} \tab \strong{OETEST} \cr
CSUBTH   \tab Center Subfield Thickness            \cr
DRSSR    \tab Diabetic Retinopathy Sev Recode Value\cr
IOP      \tab Intraocular Pressure                 \cr
VACSCORE \tab Visual Acuity Score
}
}
\author{
Gordon Miller
}
\keyword{dataset}
