% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pc}
\alias{pc}
\title{Pharmacokinetic Concentrations}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4572 rows and 20 columns.
}
\source{
\url{https://github.com/pharmaverse/pharmaversesdtm/blob/main/data/pc.rda}
}
\usage{
pc
}
\description{
A SDTM PC dataset simulated by Antonio Rodriguez Contesti
}
\author{
Antonio Rodriguez Contesti
}
\keyword{datasets}
