% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmaversesdtm-package.R
\docType{package}
\name{pharmaversesdtm-package}
\alias{pharmaversesdtm}
\alias{pharmaversesdtm-package}
\title{pharmaversesdtm: SDTM Test Data for the 'Pharmaverse' Family of Packages}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A set of Study Data Tabulation Model (SDTM) datasets from the Clinical Data Interchange Standards Consortium (CDISC) pilot project used for testing and developing Analysis Data Model (ADaM) datasets inside the pharmaverse family of packages. SDTM dataset specifications are described in the CDISC SDTM implementation guide, accessible by creating a free account on \url{https://www.cdisc.org/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pharmaverse.github.io/pharmaversesdtm/}
  \item \url{https://github.com/pharmaverse/pharmaversesdtm/}
}

}
\author{
\strong{Maintainer}: Lina Patil \email{lina.patil@cytel.com}

Authors:
\itemize{
  \item Stefan Bundfuss (\href{https://orcid.org/0009-0005-0027-1198}{ORCID})
  \item Fanny Gautier \email{fanny.gautier@cytel.com} (\href{https://orcid.org/0009-0004-3581-0131}{ORCID})
  \item Edoardo Mancini \email{edoardo.mancini@roche.com} (\href{https://orcid.org/0009-0006-4899-8641}{ORCID})
  \item Tomoyuki Namai
  \item Vinh Nguyen
  \item Vladyslav Shuliar (\href{https://orcid.org/0009-0008-2354-8999}{ORCID})
}

Other contributors:
\itemize{
  \item Cytel Inc. [copyright holder, funder]
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
  \item GlaxoSmithKline LLC [copyright holder, funder]
}

}
\keyword{internal}
