% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmaversesdtm-package.R
\docType{package}
\name{pharmaversesdtm-package}
\alias{pharmaversesdtm}
\alias{pharmaversesdtm-package}
\title{pharmaversesdtm: Test Data for the Pharmaverse Family of Packages}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A set of Study Data Tabulation Model (SDTM) datasets from the Clinical Data Interchange Standards Consortium (CDISC) pilot project used for testing and developing Analysis Data Model (ADaM) derivations inside the pharmaverse family of packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pharmaverse.github.io/pharmaversesdtm/main/}
  \item \url{https://github.com/pharmaverse/pharmaversesdtm}
}

}
\author{
\strong{Maintainer}: Edoardo Mancini \email{edoardo.mancini@roche.com}

Authors:
\itemize{
  \item Gayatri G
  \item Kangjie Zhang
  \item Pooja Kumari
  \item Stefan Bundfuss
  \item Zelos Zhu
}

}
\keyword{internal}
