% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pp}
\alias{pp}
\title{Pharmacokinetic Parameters}
\format{
An object of class \code{data.frame} with 2688 rows and 14 columns.
}
\source{
\url{https://github.com/pharmaverse/pharmaversesdtm/blob/main/data/pp.rda}
}
\usage{
pp
}
\description{
A SDTM PP dataset simulated by Antonio Rodriguez Contesti
}
\author{
Antonio Rodriguez Contesti
}
\keyword{datasets}
