% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qs_metabolic.R
\docType{data}
\name{qs_metabolic}
\alias{qs_metabolic}
\title{Questionnaire Dataset-metabolic}
\format{
A data frame with 18 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{USUBJID}{Unique Subject Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{VISIT}{Visit Name}
\item{VISITNUM}{Visit Number}
\item{VISITDY}{Planned Study Day of Visit}
\item{QSBLFL}{Baseline Flag}
\item{QSDTC}{Date/Time of Finding}
\item{QSDY}{Study Day of Finding}
\item{QSCAT}{Category for Questionnaire}
\item{QSTEST}{Questionnaire Test Name}
\item{QSTESTCD}{Questionnaire Test Short Name}
\item{QSORRES}{Result or Finding in Original Units}
\item{QSORRESU}{Original Units}
\item{QSSTRESC}{Character Result/Finding in Std Format}
\item{QSSTRESN}{Numeric Result/Finding in Standard Units}
\item{QSSTRESU}{Standard Units}
\item{QSSEQ}{Sequence Number}
}
}
\source{
Constructed by {admiralmetabolic} developers
}
\usage{
qs_metabolic
}
\description{
A SDTM QS dataset containing COEQ data (Control of Eating Questionnaire) for metabolic studies. Note that University of Leeds are the copyright holders of the Control of Eating Questionnaire (CoEQ) and the test data included within {admiralmetabolic} as well as the ADCOEQ code are for not-for-profit use only within {admiralmetabolic} and pharmaverse-related examples/documentation. Any persons or companies wanting to use the CoEQ should request a license to do so from the following \href{https://licensing.leeds.ac.uk/product/control-of-eating-questionnaire-coeq}{link}.
}
\details{
Questionnaire Dataset-metabolic

A SDTM QS dataset containing COEQ data (Control of Eating Questionnaire) for metabolic studies. Note that University of Leeds are the copyright holders of the Control of Eating Questionnaire (CoEQ) and the test data included within {admiralmetabolic} as well as the ADCOEQ code are for not-for-profit use only within {admiralmetabolic} and pharmaverse-related examples/documentation. Any persons or companies wanting to use the CoEQ should request a license to do so from the following \href{https://licensing.leeds.ac.uk/product/control-of-eating-questionnaire-coeq}{link}.
}
\keyword{dataset}
