% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/qs_ophtha.R
\docType{data}
\name{qs_ophtha}
\alias{qs_ophtha}
\title{Questionnaire for Ophthalmology}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 75922 rows and 20 columns.

A data frame with 20 columns:
\describe{
\item{STUDYID}{undocumented field}
\item{DOMAIN}{undocumented field}
\item{USUBJID}{undocumented field}
\item{QSSEQ}{undocumented field}
\item{QSTESTCD}{undocumented field}
\item{QSTEST}{undocumented field}
\item{QSCAT}{undocumented field}
\item{QSSCAT}{undocumented field}
\item{QSORRES}{undocumented field}
\item{QSORRESU}{undocumented field}
\item{QSSTRESC}{undocumented field}
\item{QSSTRESN}{undocumented field}
\item{QSSTRESU}{undocumented field}
\item{QSBLFL}{undocumented field}
\item{QSDRVFL}{undocumented field}
\item{VISITNUM}{undocumented field}
\item{VISIT}{undocumented field}
\item{VISITDY}{undocumented field}
\item{QSDTC}{undocumented field}
\item{QSDY}{undocumented field}
}
}
\source{
Constructed using \code{qs} from the \code{{pharmaversesdtm}} package

Constructed using qs from the {pharmaversesdtm} package
}
\usage{
qs_ophtha

qs_ophtha
}
\description{
An example Questionnaires SDTM dataset with ophthalmology-specific questionnaire of NEI VFQ-25

Questionnaire for Ophthalmology
}
\details{
An example Questionnaires SDTM dataset with ophthalmology-specific questionnaire of NEI VFQ-25
}
\keyword{datasets}
