% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qs_ophtha.R
\docType{data}
\name{qs_ophtha}
\alias{qs_ophtha}
\title{Questionnaire for Ophthalmology}
\format{
A data frame with 20 columns:
\describe{
\item{STUDYID}{undocumented field}
\item{DOMAIN}{undocumented field}
\item{USUBJID}{undocumented field}
\item{QSSEQ}{undocumented field}
\item{QSTESTCD}{undocumented field}
\item{QSTEST}{undocumented field}
\item{QSCAT}{undocumented field}
\item{QSSCAT}{undocumented field}
\item{QSORRES}{undocumented field}
\item{QSORRESU}{undocumented field}
\item{QSSTRESC}{undocumented field}
\item{QSSTRESN}{undocumented field}
\item{QSSTRESU}{undocumented field}
\item{QSBLFL}{undocumented field}
\item{QSDRVFL}{undocumented field}
\item{VISITNUM}{undocumented field}
\item{VISIT}{undocumented field}
\item{VISITDY}{undocumented field}
\item{QSDTC}{undocumented field}
\item{QSDY}{undocumented field}
}
}
\source{
Constructed using qs from the {pharmaversesdtm} package
}
\usage{
qs_ophtha
}
\description{
An example Questionnaires SDTM dataset with ophthalmology-specific questionnaire of NEI VFQ-25
}
\details{
Questionnaire for Ophthalmology

An example Questionnaires SDTM dataset with ophthalmology-specific questionnaire of NEI VFQ-25

Contains a list of 29 unique test codes and names:
\describe{
\item{VFQ1}{Your Overall Health Is}
\item{VFQ18}{Driving in Difficult Conditions}
\item{VFQ2}{Eyesight Using Both Eyes Is}
\item{VFQ17}{Difficulty Driving at Night}
\item{VFQ29}{Difficulty Seeing Programs on TV}
\item{VFQ16}{Difficulty Driving During Daytime}
\item{VFQ12}{Difficulty Picking Out Own Clothes}
\item{VFQ28}{Difficulty Taking Part in Sports}
\item{VFQ9}{Difficulty Going Down Step at Night}
\item{VFQ26}{Difficulty Shaving or Styling Hair}
\item{VFQ23}{Worry I'll Do Embarrassing Things}
\item{VFQ8}{Difficulty Reading Street Signs}
\item{VFQ20}{I Stay Home Most of the Time}
\item{VFQ6}{Difficulty Doing Work/Hobbies}
\item{VFQ3}{How Often You Worry About Eyesight}
\item{VFQ21}{I Feel Frustrated a Lot of the Time}
\item{VFQ7}{Difficulty Finding on Crowded Shelf}
\item{VFQ14}{Difficulty Going Out to See Movies}
\item{VFQ15}{Are You Currently Driving}
\item{VFQ13}{Difficulty Visiting With People}
\item{VFQ4}{How Often Pain in and Around Eyes}
\item{VFQ25}{Difficulty Figure Out Bill Accuracy}
\item{VFQ11}{Difficulty Seeing How People React}
\item{VFQ10}{Difficulty Noticing Objects to Side}
\item{VFQ5}{Difficulty Reading Newspapers}
\item{VFQ22}{I Need a Lot of Help From Others}
\item{VFQ19}{Eye Pain Keep You From Doing What You Like}
\item{VFQ27}{Difficulty Recognizing People}
\item{VFQ24}{Difficulty Reading Small Print}
}
}
\keyword{dataset}
