% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qs_ophtha.R
\docType{data}
\name{qs_ophtha}
\alias{qs_ophtha}
\title{Questionnaire for Ophthalmology}
\format{
A data frame with 20 columns:
\describe{
\item{STUDYID}{undocumented field}
\item{DOMAIN}{undocumented field}
\item{USUBJID}{undocumented field}
\item{QSSEQ}{undocumented field}
\item{QSTESTCD}{undocumented field}
\item{QSTEST}{undocumented field}
\item{QSCAT}{undocumented field}
\item{QSSCAT}{undocumented field}
\item{QSORRES}{undocumented field}
\item{QSORRESU}{undocumented field}
\item{QSSTRESC}{undocumented field}
\item{QSSTRESN}{undocumented field}
\item{QSSTRESU}{undocumented field}
\item{QSBLFL}{undocumented field}
\item{QSDRVFL}{undocumented field}
\item{VISITNUM}{undocumented field}
\item{VISIT}{undocumented field}
\item{VISITDY}{undocumented field}
\item{QSDTC}{undocumented field}
\item{QSDY}{undocumented field}
}
}
\source{
Constructed using qs from the {pharmaversesdtm} package
}
\usage{
qs_ophtha
}
\description{
An example Questionnaires SDTM dataset with ophthalmology-specific questionnaire of NEI VFQ-25
}
\details{
Questionnaire for Ophthalmology

An example Questionnaires SDTM dataset with ophthalmology-specific questionnaire of NEI VFQ-25
}
\keyword{dataset}
