% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qs_ophtha}
\alias{qs_ophtha}
\title{Questionnaire for Ophthalmology}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 75922 rows and 20 columns.
}
\source{
Constructed using \code{qs} from the \code{{pharmaversesdtm}} package
}
\usage{
qs_ophtha
}
\description{
An example Questionnaires SDTM dataset with ophthalmology-specific questionnaire of NEI VFQ-25
}
\keyword{datasets}
