% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rs_onco}
\alias{rs_onco}
\title{Disease Response Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5808 rows and 19 columns.
}
\usage{
rs_onco
}
\description{
A SDTM RS dataset simulated by Gopi Vegesna
}
\author{
Gopi Vegesna
}
\keyword{datasets}
