% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_onco_ca125.R
\docType{data}
\name{rs_onco_ca125}
\alias{rs_onco_ca125}
\title{No label available}
\format{
A data frame with 13 columns:
\describe{
\item{STUDYID}{undocumented field}
\item{DOMAIN}{undocumented field}
\item{USUBJID}{undocumented field}
\item{RSSEQ}{undocumented field}
\item{RSTESTCD}{undocumented field}
\item{RSTEST}{undocumented field}
\item{RSCAT}{undocumented field}
\item{RSORRES}{undocumented field}
\item{RSSTRESC}{undocumented field}
\item{RSEVAL}{undocumented field}
\item{VISITNUM}{undocumented field}
\item{VISIT}{undocumented field}
\item{RSDTC}{undocumented field}
}
}
\source{
No source available
}
\usage{
rs_onco_ca125
}
\description{
dataset
}
\details{
No label available

No description available
}
\keyword{dataset}
