% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_onco_ca125.R
\docType{data}
\name{rs_onco_ca125}
\alias{rs_onco_ca125}
\title{Disease Response (GCIG)}
\format{
A data frame with 13 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{RSSEQ}{Sequence Number}
\item{RSTESTCD}{Assessment Short Name}
\item{RSTEST}{Assessment Name}
\item{RSCAT}{Category for Assessment}
\item{RSORRES}{Result or Finding in Original Units}
\item{RSSTRESC}{Character Result/Finding in Std Format}
\item{RSEVAL}{Evaluator}
\item{VISITNUM}{Visit Number}
\item{VISIT}{Visit Name}
\item{RSDTC}{Date/Time of Assessment}
}
}
\source{
Generated dataset (\href{https://github.com/pharmaverse/pharmaversesdtm/blob/main/data-raw/rs_supprs_onco_ca125.R}{rs_supprs_onco_ca125.R})
}
\usage{
rs_onco_ca125
}
\description{
A SDTM RS dataset using GCIG criteria. The dataset contains just a few patients. It is intended for vignettes and examples of ADaM dataset creation.
}
\details{
Disease Response (GCIG)

A SDTM RS dataset using GCIG criteria. The dataset contains just a few patients. It is intended for vignettes and examples of ADaM dataset creation.

Contains a set of 1 unique Test Short Name and Test Name: \tabular{ll}{
\strong{RSTESTCD} \tab \strong{RSTEST} \cr
OVRLRESP \tab Overall Response
}
}
\author{
Vinh Nguyen
}
\keyword{dataset}
