% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_onco_ca125.R
\docType{data}
\name{rs_onco_ca125}
\alias{rs_onco_ca125}
\title{Disease Response (GCIG)}
\format{
A data frame with 13 columns:
\describe{
\item{STUDYID}{undocumented field}
\item{DOMAIN}{undocumented field}
\item{USUBJID}{undocumented field}
\item{RSSEQ}{undocumented field}
\item{RSTESTCD}{undocumented field}
\item{RSTEST}{undocumented field}
\item{RSCAT}{undocumented field}
\item{RSORRES}{undocumented field}
\item{RSSTRESC}{undocumented field}
\item{RSEVAL}{undocumented field}
\item{VISITNUM}{undocumented field}
\item{VISIT}{undocumented field}
\item{RSDTC}{undocumented field}
}
}
\source{
Generated dataset (\href{https://github.com/pharmaverse/pharmaversesdtm/blob/main/data-raw/rs_supprs_onco_ca125.R}{rs_supprs_onco_ca125.R})
}
\usage{
rs_onco_ca125
}
\description{
A SDTM RS dataset using GCIG criteria. The dataset contains just a few patients. It is intended for vignettes and examples of ADaM dataset creation.
}
\details{
Disease Response (GCIG)

A SDTM RS dataset using GCIG criteria. The dataset contains just a few patients. It is intended for vignettes and examples of ADaM dataset creation.
}
\author{
Vinh Nguyen
}
\keyword{dataset}
