% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rs_onco_imwg}
\alias{rs_onco_imwg}
\title{Disease Response (IMWG)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 65 rows and 17 columns.
}
\usage{
rs_onco_imwg
}
\description{
A SDTM RS dataset using IMWG criteria. The dataset contains just a few patients.
It is intended for vignettes and examples of ADaM dataset creation.
}
\author{
Vinh Nguyen
}
\keyword{datasets}
