% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_onco_irecist.R
\docType{data}
\name{rs_onco_irecist}
\alias{rs_onco_irecist}
\title{Disease Response (iRECIST) for Oncology}
\format{
A data frame with 19 columns:
\describe{
\item{STUDYID}{undocumented field}
\item{DOMAIN}{undocumented field}
\item{USUBJID}{undocumented field}
\item{RSSEQ}{undocumented field}
\item{RSLNKGRP}{undocumented field}
\item{RSTESTCD}{undocumented field}
\item{RSTEST}{undocumented field}
\item{RSCAT}{undocumented field}
\item{RSORRES}{undocumented field}
\item{RSSTRESC}{undocumented field}
\item{RSSTAT}{undocumented field}
\item{RSREASND}{undocumented field}
\item{RSEVAL}{undocumented field}
\item{RSEVALID}{undocumented field}
\item{RSACPTFL}{undocumented field}
\item{VISITNUM}{undocumented field}
\item{VISIT}{undocumented field}
\item{RSDTC}{undocumented field}
\item{RSDY}{undocumented field}
}
}
\source{
Generated dataset.
}
\usage{
rs_onco_irecist
}
\description{
A SDTM RS dataset using iRECIST intended for examples of ADaM dataset creation
}
\details{
Disease Response (iRECIST) for Oncology

A SDTM RS dataset using iRECIST intended for examples of ADaM dataset creation

Contains a list of 6 unique Test Short Names and Test Names: \tabular{ll}{
\strong{RSTESTCD} \tab \strong{RSTEST} \cr
IRECLIND \tab Last iRECIST Assessment Indicator\cr
NEWLIND  \tab New Lesion Indicator             \cr
NEWLPROG \tab New Lesion Progression           \cr
NTRGRESP \tab Non-target Response              \cr
OVRLRESP \tab Overall Response                 \cr
TRGRESP  \tab Target Response
}
}
\author{
Rohan Thampi
}
\keyword{dataset}
