% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rs_onco_irecist}
\alias{rs_onco_irecist}
\title{Disease Response Dataset (iRECIST)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 376 rows and 19 columns.
}
\usage{
rs_onco_irecist
}
\description{
A SDTM RS dataset using iRECIST. The dataset contains just a few patients.
It is intended for vignettes and examples of ADaM dataset creation.
}
\author{
Rohan Thampi
}
\keyword{datasets}
