% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_onco_irecist.R
\docType{data}
\name{rs_onco_irecist}
\alias{rs_onco_irecist}
\title{Disease Response (iRECIST) for Oncology}
\format{
A data frame with 19 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{RSSEQ}{Sequence Number}
\item{RSLNKGRP}{Link Group ID}
\item{RSTESTCD}{Assessment Short Name}
\item{RSTEST}{Assessment Name}
\item{RSCAT}{Category for Assessment}
\item{RSORRES}{Result or Finding in Original Units}
\item{RSSTRESC}{Character Result/Finding in Std Format}
\item{RSSTAT}{Completion Status}
\item{RSREASND}{Reason Not Done}
\item{RSEVAL}{Evaluator}
\item{RSEVALID}{Evaluator Identifier}
\item{RSACPTFL}{Accepted Record Flag}
\item{VISITNUM}{Visit Number}
\item{VISIT}{Visit Name}
\item{RSDTC}{Date/Time of Assessment}
\item{RSDY}{Study Day of Assessment}
}
}
\source{
Generated dataset.
}
\usage{
rs_onco_irecist
}
\description{
A SDTM RS dataset using iRECIST intended for examples of ADaM dataset creation
}
\details{
Disease Response (iRECIST) for Oncology

A SDTM RS dataset using iRECIST intended for examples of ADaM dataset creation

Contains a set of 6 unique Test Short Names and Test Names: \tabular{ll}{
\strong{RSTESTCD} \tab \strong{RSTEST} \cr
IRECLIND \tab Last iRECIST Assessment Indicator\cr
NEWLIND  \tab New Lesion Indicator             \cr
NEWLPROG \tab New Lesion Progression           \cr
NTRGRESP \tab Non-target Response              \cr
OVRLRESP \tab Overall Response                 \cr
TRGRESP  \tab Target Response
}
}
\author{
Rohan Thampi
}
\keyword{dataset}
