% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_onco_recist.R
\docType{data}
\name{rs_onco_recist}
\alias{rs_onco_recist}
\title{Disease Response (RECIST 1.1) for Oncology}
\format{
A data frame with 14 columns:
\describe{
\item{DOMAIN}{Domain Abbreviation}
\item{STUDYID}{Study Identifier}
\item{USUBJID}{Unique Subject Identifier}
\item{VISITNUM}{Visit Number}
\item{VISIT}{Visit Name}
\item{RSTESTCD}{Assessment Short Name}
\item{RSTEST}{Assessment Name}
\item{RSORRES}{Result or Finding in Original Units}
\item{RSSTRESC}{Character Result/Finding in Std Format}
\item{RSEVAL}{Evaluator}
\item{RSEVALID}{Evaluator Identifier}
\item{RSACPTFL}{Accepted Record Flag}
\item{RSDTC}{Date/Time of Assessment}
\item{RSSEQ}{Sequence Number}
}
}
\source{
Generated dataset.
}
\usage{
rs_onco_recist
}
\description{
A SDTM RS dataset using RECIST 1.1 intended for examples of ADaM dataset creation
}
\details{
Disease Response (RECIST 1.1) for Oncology

A SDTM RS dataset using RECIST 1.1 intended for examples of ADaM dataset creation

Contains a set of 1 unique Test Short Name and Test Name: \tabular{ll}{
\strong{RSTESTCD} \tab \strong{RSTEST} \cr
OVRLRESP \tab Overall Response
}
}
\author{
Stefan Bundfuss
}
\keyword{dataset}
