% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_onco_recist.R
\docType{data}
\name{rs_onco_recist}
\alias{rs_onco_recist}
\title{Disease Response (RECIST 1.1) for Oncology}
\format{
A data frame with 14 columns:
\describe{
\item{DOMAIN}{undocumented field}
\item{STUDYID}{undocumented field}
\item{USUBJID}{undocumented field}
\item{VISITNUM}{undocumented field}
\item{VISIT}{undocumented field}
\item{RSTESTCD}{undocumented field}
\item{RSTEST}{undocumented field}
\item{RSORRES}{undocumented field}
\item{RSSTRESC}{undocumented field}
\item{RSEVAL}{undocumented field}
\item{RSEVALID}{undocumented field}
\item{RSACPTFL}{undocumented field}
\item{RSDTC}{undocumented field}
\item{RSSEQ}{undocumented field}
}
}
\source{
Generated dataset.
}
\usage{
rs_onco_recist
}
\description{
A SDTM RS dataset using RECIST 1.1 intended for examples of ADaM dataset creation
}
\details{
Disease Response (RECIST 1.1) for Oncology

A SDTM RS dataset using RECIST 1.1 intended for examples of ADaM dataset creation
}
\author{
Stefan Bundfuss
}
\keyword{dataset}
