% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rs_onco_recist}
\alias{rs_onco_recist}
\title{Disease Response Dataset (RECIST 1.1)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 66 rows and 14 columns.
}
\source{
The dataset is derived from \code{tr_onco_recist}.
}
\usage{
rs_onco_recist
}
\description{
A SDTM RS dataset using RECIST 1.1. The dataset contains just a few patients.
It is intended for vignettes and examples of ADaM dataset creation.
}
\author{
Stefan Bundfuss
}
\keyword{datasets}
