% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdg_db.R
\docType{data}
\name{sdg_db}
\alias{sdg_db}
\title{SDG}
\format{
A data frame with 5 columns:
\describe{
\item{termchar}{undocumented field}
\item{sdg_name}{undocumented field}
\item{sdg_id}{undocumented field}
\item{termvar}{undocumented field}
\item{version}{undocumented field}
}
}
\source{
\if{html}{\out{<a href="https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/sv.xpt?raw=true" target="_blank">}}Access the source of the SDG dataset.\if{html}{\out{</a>}}
}
\usage{
sdg_db
}
\description{
An example SDG dataset
}
\details{
SDG

An example SDG dataset
}
\keyword{dataset}
