% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sdg_db}
\alias{sdg_db}
\title{SDG}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16 rows and 5 columns.
}
\usage{
sdg_db
}
\description{
An example SDG dataset
}
\keyword{datasets}
