% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smq_db.R
\docType{data}
\name{smq_db}
\alias{smq_db}
\title{Standardized MedDRA Queries}
\format{
A data frame with 6 columns:
\describe{
\item{termchar}{Reported Term}
\item{scope}{Scope}
\item{smq_name}{Name}
\item{smq_id}{Name ID}
\item{version}{Version}
\item{termvar}{Variable}
}
}
\source{
Generated dataset.
}
\usage{
smq_db
}
\description{
An example SMQ dataset
}
\details{
Standardized MedDRA Queries

An example SMQ dataset
}
\keyword{dataset}
