% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smq_db}
\alias{smq_db}
\title{SMQ Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 44 rows and 6 columns.
}
\usage{
smq_db
}
\description{
An example SMQ dataset
}
\keyword{datasets}
