% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppae.R
\docType{data}
\name{suppae}
\alias{suppae}
\title{Supplemental Adverse Events}
\format{
A data frame with 10 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{RDOMAIN}{Related Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{IDVAR}{Identifying Variable}
\item{IDVARVAL}{Identifying Variable Value}
\item{QNAM}{Qualifier Variable Name}
\item{QLABEL}{Qualifier Variable Label}
\item{QVAL}{Data Value}
\item{QORIG}{Origin}
\item{QEVAL}{Evaluator}
}
}
\source{
\href{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/suppds.xpt?raw=true}{Access the source of the Supplemental Adverse Events dataset.}
}
\usage{
suppae
}
\description{
A SDTM SUPPAE dataset from the CDISC pilot project
}
\details{
Supplemental Adverse Events

A SDTM SUPPAE dataset from the CDISC pilot project
}
\keyword{dataset}
