% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppce_vaccine.R
\docType{data}
\name{suppce_vaccine}
\alias{suppce_vaccine}
\title{Supplemental Qualifiers for Clinical Events for Vaccine}
\format{
A data frame with 9 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{USUBJID}{Unique Subject Identifier}
\item{RDOMAIN}{Related Domain Abbreviation}
\item{IDVAR}{Identifying Variable}
\item{IDVARVAL}{Identifying Variable Value}
\item{QNAM}{Qualifier Variable Name}
\item{QVAL}{Data Value}
\item{QLABEL}{Qualifier Variable Label}
\item{QORIG}{Origin}
}
}
\source{
Constructed by {admiralvaccine} developers
}
\usage{
suppce_vaccine
}
\description{
An example SDTM SUPPCE for vaccine studies
}
\details{
Supplemental Qualifiers for Clinical Events for Vaccine

An example SDTM SUPPCE for vaccine studies
}
\keyword{dataset}
