% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppdm.R
\docType{data}
\name{suppdm}
\alias{suppdm}
\title{Supplemental Demography}
\format{
A data frame with 10 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{RDOMAIN}{Related Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{IDVAR}{Identifying Variable}
\item{IDVARVAL}{Identifying Variable Value}
\item{QNAM}{Qualifier Variable Name}
\item{QLABEL}{Qualifier Variable Label}
\item{QVAL}{Data Value}
\item{QORIG}{Origin}
\item{QEVAL}{Evaluator}
}
}
\source{
Generated dataset.
}
\usage{
suppdm
}
\description{
A SDTM SUPPDM dataset from the CDISC pilot project
}
\details{
Supplemental Demography

A SDTM SUPPDM dataset from the CDISC pilot project
}
\keyword{dataset}
