% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{suppdm_vaccine}
\alias{suppdm_vaccine}
\title{Supplemental Qualifiers for Demographics for Vaccine}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2 rows and 9 columns.
}
\source{
Constructed by \code{{admiralvaccine}} developers
}
\usage{
suppdm_vaccine
}
\description{
An example SDTM SUPPDM dataset for vaccine studies
}
\keyword{datasets}
