% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppds.R
\docType{data}
\name{suppds}
\alias{suppds}
\title{Supplemental Disposition}
\format{
A data frame with 9 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{RDOMAIN}{Related Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{IDVAR}{Identifying Variable}
\item{IDVARVAL}{Identifying Variable Value}
\item{QNAM}{Qualifier Variable Name}
\item{QLABEL}{Qualifier Variable Label}
\item{QVAL}{Data Value}
\item{QORIG}{Origin}
}
}
\source{
\href{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/suppdm.xpt?raw=true}{Access the source of the Supplemental Disposition dataset.}
}
\usage{
suppds
}
\description{
A SDTM SUPPDS dataset from the CDISC pilot project
}
\details{
Supplemental Disposition

A SDTM SUPPDS dataset from the CDISC pilot project
}
\keyword{dataset}
