% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{suppex_vaccine}
\alias{suppex_vaccine}
\title{Supplemental Qualifiers for Exposures Dataset for Vaccine Studies}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 9 columns.
}
\source{
\url{https://github.com/pharmaverse/admiralvaccine/blob/main/data/vx_suppex.rda} # nolint
}
\usage{
suppex_vaccine
}
\description{
An example SDTM SUPPEX dataset for vaccine studies
}
\keyword{datasets}
