% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{suppface_vaccine}
\alias{suppface_vaccine}
\title{Supplemental Qualifiers for Findings About for Clinical Events for Vaccine}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 9 columns.
}
\source{
Constructed by \code{{admiralvaccine}} developers
}
\usage{
suppface_vaccine
}
\description{
An example SDTM SUPPFACE dataset for vaccine studies
}
\keyword{datasets}
