% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{suppis_vaccine}
\alias{suppis_vaccine}
\title{Supplemental Qualifiers for Immunogenicity Specimen Assessments for Vaccine}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16 rows and 10 columns.
}
\source{
Constructed by \code{{admiralvaccine}} developers
}
\usage{
suppis_vaccine
}
\description{
An example SDTM SUPPIS dataset for vaccine studies
}
\keyword{datasets}
