% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppis_vaccine.R
\docType{data}
\name{suppis_vaccine}
\alias{suppis_vaccine}
\title{Supplemental Qualifiers for Immunogenicity Specimen Assessments for Vaccine}
\format{
A data frame with 10 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{RDOMAIN}{Related Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{IDVAR}{Identifying Variable}
\item{IDVARVAL}{Identifying Variable Value}
\item{QNAM}{Qualifier Variable Name}
\item{QLABEL}{Qualifier Variable Label}
\item{QVAL}{Data Value}
\item{QORIG}{Origin}
\item{QEVAL}{Evaluator}
}
}
\source{
Constructed by {admiralpeds} developers
}
\usage{
suppis_vaccine
}
\description{
An example SDTM SUPPIS dataset for vaccine studies
}
\details{
Supplemental Qualifiers for Immunogenicity Specimen Assessments for Vaccine

An example SDTM SUPPIS dataset for vaccine studies
}
\keyword{dataset}
