% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{supprs_onco_ca125}
\alias{supprs_onco_ca125}
\title{Supplemental Qualifiers for RS_ONCO_CA125}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 49 rows and 9 columns.
}
\usage{
supprs_onco_ca125
}
\description{
A SDTM supplemental RS dataset using GCIG criteria. It is intended to be used
together with \code{rs_onco_ca125}.
}
\author{
Vinh Nguyen
}
\keyword{datasets}
