% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supprs_onco_ca125.R
\docType{data}
\name{supprs_onco_ca125}
\alias{supprs_onco_ca125}
\title{Supplemental Qualifiers for RS_ONCO_CA125}
\format{
A data frame with 9 columns:
\describe{
\item{STUDYID}{undocumented field}
\item{RDOMAIN}{undocumented field}
\item{USUBJID}{undocumented field}
\item{IDVAR}{undocumented field}
\item{IDVARVAL}{undocumented field}
\item{QNAM}{undocumented field}
\item{QLABEL}{undocumented field}
\item{QVAL}{undocumented field}
\item{QORIG}{undocumented field}
}
}
\source{
Generated dataset (\href{https://github.com/pharmaverse/pharmaversesdtm/blob/main/data-raw/rs_supprs_onco_ca125.R}{rs_supprs_onco_ca125.R})
}
\usage{
supprs_onco_ca125
}
\description{
A SDTM supplemental RS dataset using GCIG criteria. It is intended to be used together with \code{rs_onco_ca125}.
}
\details{
Supplemental Qualifiers for RS_ONCO_CA125

A SDTM supplemental RS dataset using GCIG criteria. It is intended to be used together with \code{rs_onco_ca125}.
}
\author{
Vinh Nguyen
}
\keyword{dataset}
