% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supprs_onco_imwg.R
\docType{data}
\name{supprs_onco_imwg}
\alias{supprs_onco_imwg}
\title{Supplemental Qualifiers for RS_ONCO_IMWG}
\format{
A data frame with 9 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{RDOMAIN}{Related Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{IDVAR}{Identifying Variable}
\item{IDVARVAL}{Identifying Variable Value}
\item{QNAM}{Qualifier Variable Name}
\item{QLABEL}{Qualifier Variable Label}
\item{QVAL}{Data Value}
\item{QORIG}{Origin}
}
}
\source{
\href{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/suppae.xpt?raw=true}{Access the source of the Supplemental Qualifiers for RS_ONCO_IMWG dataset.}
}
\usage{
supprs_onco_imwg
}
\description{
A SDTM supplemental RS dataset using IMWG criteria intended to be used with \code{rs_onco_imwg}
}
\details{
Supplemental Qualifiers for RS_ONCO_IMWG

A SDTM supplemental RS dataset using IMWG criteria intended to be used with \code{rs_onco_imwg}
}
\author{
Vinh Nguyen
}
\keyword{dataset}
