% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supprs_onco_imwg.R
\docType{data}
\name{supprs_onco_imwg}
\alias{supprs_onco_imwg}
\title{Supplemental Qualifiers for RS_ONCO_IMWG}
\format{
A data frame with 9 columns:
\describe{
\item{STUDYID}{undocumented field}
\item{RDOMAIN}{undocumented field}
\item{USUBJID}{undocumented field}
\item{IDVAR}{undocumented field}
\item{IDVARVAL}{undocumented field}
\item{QNAM}{undocumented field}
\item{QLABEL}{undocumented field}
\item{QVAL}{undocumented field}
\item{QORIG}{undocumented field}
}
}
\source{
\href{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/suppae.xpt?raw=true}{Access the source of the Supplemental Qualifiers for RS_ONCO_IMWG dataset.}
}
\usage{
supprs_onco_imwg
}
\description{
A SDTM supplemental RS dataset using IMWG criteria intended to be used with \code{rs_onco_imwg}
}
\details{
Supplemental Qualifiers for RS_ONCO_IMWG

A SDTM supplemental RS dataset using IMWG criteria intended to be used with \code{rs_onco_imwg}
}
\author{
Vinh Nguyen
}
\keyword{dataset}
