% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{supprs_onco_imwg}
\alias{supprs_onco_imwg}
\title{Supplemental Qualifiers for RS_ONCO_IMWG}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 19 rows and 9 columns.
}
\usage{
supprs_onco_imwg
}
\description{
A SDTM supplemental RS dataset using IMWG criteria. It is intended to be used
together with \code{rs_onco_imwg}.
}
\author{
Vinh Nguyen
}
\keyword{datasets}
