% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supptr_onco.R
\docType{data}
\name{supptr_onco}
\alias{supptr_onco}
\title{Supplemental Tumor Results for Oncology}
\format{
A data frame with 9 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{RDOMAIN}{Related Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{IDVAR}{Identifying Variable}
\item{IDVARVAL}{Identifying Variable Value}
\item{QNAM}{Qualifier Variable Name}
\item{QLABEL}{Qualifier Variable Label}
\item{QVAL}{Data Value}
\item{QORIG}{Origin}
}
}
\source{
\if{html}{\out{<a href="https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/ts.xpt?raw=true" target="_blank">}}Access the source of the Supplemental Tumor Results for Oncology dataset.\if{html}{\out{</a>}}
}
\usage{
supptr_onco
}
\description{
dataset
}
\details{
Supplemental Tumor Results for Oncology

A SDTM SUPPTR dataset simulated by Gopi Vegesna
}
\author{
Gopi Vegesna
}
\keyword{dataset}
