% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{supptr_onco}
\alias{supptr_onco}
\title{Supplemental Tumor Results for Oncology}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 55995 rows and 9 columns.
}
\usage{
supptr_onco
}
\description{
A SDTM SUPPTR dataset simulated by Gopi Vegesna
}
\author{
Gopi Vegesna
}
\keyword{datasets}
