% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv.R
\docType{data}
\name{sv}
\alias{sv}
\title{Subject Visits}
\format{
A data frame with 8 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{VISITNUM}{Visit Number}
\item{VISIT}{Visit Name}
\item{VISITDY}{Planned Study Day of Visit}
\item{SVSTDTC}{Start Date/Time of Visit}
\item{SVENDTC}{End Date/Time of Visit}
}
}
\source{
Constructed by {admiralvaccine} developers
}
\usage{
sv
}
\description{
A SDTM SV dataset from the CDISC pilot project with corrected duplicate observation
}
\details{
Subject Visits

A SDTM SV dataset from the CDISC pilot project with corrected duplicate observation
}
\keyword{dataset}
