% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_onco_recist.R
\docType{data}
\name{tr_onco_recist}
\alias{tr_onco_recist}
\title{Tumor Results (RECIST 1.1) for Oncology}
\format{
A data frame with 19 columns:
\describe{
\item{DOMAIN}{undocumented field}
\item{STUDYID}{undocumented field}
\item{USUBJID}{undocumented field}
\item{TRGRPID}{undocumented field}
\item{TRLNKID}{undocumented field}
\item{TRTESTCD}{undocumented field}
\item{TRTEST}{undocumented field}
\item{TRORRES}{undocumented field}
\item{TRORRESU}{undocumented field}
\item{TRSTRESC}{undocumented field}
\item{TRSTRESN}{undocumented field}
\item{TRSTRESU}{undocumented field}
\item{VISITNUM}{undocumented field}
\item{VISIT}{undocumented field}
\item{TREVAL}{undocumented field}
\item{TREVALID}{undocumented field}
\item{TRACPTFL}{undocumented field}
\item{TRDTC}{undocumented field}
\item{TRSEQ}{undocumented field}
}
}
\source{
Generated dataset.
}
\usage{
tr_onco_recist
}
\description{
A SDTM TR dataset using RECIST 1.1 intended for examples of ADaM dataset creation
}
\details{
Tumor Results (RECIST 1.1) for Oncology

A SDTM TR dataset using RECIST 1.1 intended for examples of ADaM dataset creation
}
\author{
Stefan Bundfuss
}
\keyword{dataset}
