% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\docType{data}
\name{ts}
\alias{ts}
\title{Trial Design}
\format{
A data frame with 6 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{TSSEQ}{Sequence Number}
\item{TSPARMCD}{Trial Summary Parameter Short Name}
\item{TSPARM}{Trial Summary Parameter}
\item{TSVAL}{Parameter Value}
}
}
\source{
\href{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/vs.xpt?raw=true}{Access the source of the Trial Design dataset.}
}
\usage{
ts
}
\description{
A SDTM TS dataset from the CDISC pilot project
}
\details{
Trial Design

A SDTM TS dataset from the CDISC pilot project
}
\keyword{dataset}
