% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tu_onco.R
\docType{data}
\name{tu_onco}
\alias{tu_onco}
\title{Tumor Identification for Oncology}
\format{
A data frame with 18 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{TUSEQ}{Sequence Number}
\item{TULNKID}{Link ID}
\item{TUTESTCD}{Tumor Identification Short Name}
\item{TUTEST}{Tumor Identification Test Name}
\item{TUORRES}{Tumor Identification Result}
\item{TUSTRESC}{Tumor Identification Result Std. Format}
\item{TULOC}{Location of the Tumor}
\item{TUMETHOD}{Method of Identification}
\item{TUEVAL}{Evaluator}
\item{TUEVALID}{Evaluator Identifier}
\item{TUACPTFL}{Accepted Record Flag}
\item{VISITNUM}{Visit Number}
\item{VISIT}{Visit Name}
\item{TUDTC}{Date/Time of Tumor Identification}
\item{TUDY}{Study Day of Tumor Identification}
}
}
\source{
Generated dataset.
}
\usage{
tu_onco
}
\description{
A SDTM TU dataset simulated by Gopi Vegesna
}
\details{
Tumor Identification for Oncology

A SDTM TU dataset simulated by Gopi Vegesna
}
\author{
Gopi Vegesna
}
\keyword{dataset}
