% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tu_onco}
\alias{tu_onco}
\title{Tumor Identification for Oncology}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7734 rows and 18 columns.
}
\usage{
tu_onco
}
\description{
A SDTM TU dataset simulated by Gopi Vegesna
}
\author{
Gopi Vegesna
}
\keyword{datasets}
