% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tu_onco_recist}
\alias{tu_onco_recist}
\title{Tumor Identification (RECIST 1.1) for Oncology}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 75 rows and 16 columns.
}
\usage{
tu_onco_recist
}
\description{
A SDTM TU dataset using RECIST 1.1. The dataset contains just a few patients.
It is intended for vignettes and examples of ADaM dataset creation.
}
\author{
Stefan Bundfuss
}
\keyword{datasets}
