% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vs_vaccine}
\alias{vs_vaccine}
\title{Vital Signs for Vaccine}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 28 rows and 23 columns.
}
\source{
Constructed by \code{{admiralvaccine}} developers
}
\usage{
vs_vaccine
}
\description{
An example SDTM VS dataset for vaccine studies
}
\keyword{datasets}
